% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addContour}
\alias{addContour}
\title{addContour}
\usage{
addContour(map, val, col = "blue", super = TRUE)
}
\arguments{
\item{map}{object returned by function genMap}

\item{val}{quantile value vector}

\item{col}{color parameter}

\item{super}{if TRUE add to existing plot lines coresponding to contour, if FALSE plot boundary and add lines}
}
\value{
void
}
\description{
addContour
}
\details{
add contour lines to plot
}
\examples{
data(mapTest)
addContour(mapTest,c(5,7),super=FALSE)
}
