% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingZone.R
\name{buffToValid}
\alias{buffToValid}
\title{buffToValid}
\usage{
buffToValid(zone)
}
\arguments{
\item{zone}{a SpatialPolygon}
}
\value{
a new valid zone
}
\description{
buffToValid
}
\details{
function that check if a zone has a valid geometry, if not , makes zone valid by using gBuffer(width = 0,...)
}
