% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calcCad}
\alias{calcCad}
\title{calcCad}
\usage{
calcCad(cad)
}
\arguments{
\item{cad}{xxxx}
}
\value{
a SpatialPolygons object
}
\description{
calcCad
}
\details{
description, a paragraph
}
\examples{
# not run
}
