% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalNei.R
\name{calcCritNarrow}
\alias{calcCritNarrow}
\title{detection of narrow zones (ratio area/perimeter^2)}
\usage{
calcCritNarrow(zonePolygone)
}
\arguments{
\item{zonePolygone}{xxxx}
}
\value{
a ?
}
\description{
detection of narrow zones (ratio area/perimeter^2)
}
\details{
description, a paragraph
}
\examples{
# not run
}
