% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calcDCrit}
\alias{calcDCrit}
\title{calcDCrit}
\usage{
calcDCrit(Z, map, optiCrit, pErr = 0.9, simplitol = 0.001)
}
\arguments{
\item{Z}{xxxx}

\item{map}{xxxx}

\item{optiCrit}{xxxx}

\item{pErr}{xxxx, default 0.9}

\item{simplitol}{xxxx, default 0.001}
}
\value{
a list
}
\description{
calcDCrit
}
\details{
description, a paragraph
}
\examples{
# not run
}
