% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calCrit.R
\name{calCrit}
\alias{calCrit}
\title{calCrit}
\usage{
calCrit(matDistanceCorr, zoneNModif, optiCrit = 2)
}
\arguments{
\item{matDistanceCorr}{xxxx}

\item{zoneNModif}{xxxx}

\item{optiCrit}{xxxx}
}
\value{
a ?
}
\description{
calCrit
}
\details{
description, a paragraph
}
\examples{
# not run
}
