% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit2}
\alias{calCrit2}
\title{calCrit2}
\usage{
calCrit2(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCrit2
}
\details{
description, a paragraph
}
\examples{
# not run
}
