% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit3}
\alias{calCrit3}
\title{calCrit3}
\usage{
calCrit3(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{zone distance matrix resulting from a call to calDistance}

\item{zoneNModif}{matrix of zone neigbors with FALSE on the diagonal}
}
\value{
a numerical value
}
\description{
calCrit3
}
\details{
computes a quality criterion equal to min(mean(dij^2/sqrt(dii^2*dij^2))) 
see also  \code{\link{calCrit1}}, \code{\link{calCrit2}}, \code{\link{calCrit4}}, \code{\link{calCrit5}}, \code{\link{calCritMinMean}} for other criteria
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
resD = calDistance(typedist=1,mapTest$krigData,K$listZonePoint,K$zoneN,
       mapTest$krigSurfVoronoi,K$meanZone,pErr=0.9)
calCrit3(resD$matDistanceCorr,K$zoneNModif)
# not run
}
