% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit4}
\alias{calCrit4}
\title{calCrit4}
\usage{
calCrit4(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCrit4
}
\details{
description, a paragraph
}
\examples{
# not run
}
