% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit5}
\alias{calCrit5}
\title{calCrit5}
\usage{
calCrit5(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCrit5
}
\details{
description, a paragraph
}
\examples{
# not run
}
