% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit7}
\alias{calCrit7}
\title{calCrit7}
\usage{
calCrit7(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{zone distance matrix resulting from a call to calDistance}

\item{zoneNModif}{matrix of zone neigbors with FALSE on the diagonal}
}
\value{
a numerical value
}
\description{
calCrit7
}
\details{
computes a quality criterion equal to mean(2*mean(dij/(dii+djj)))
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
resD = calDistance(typedist=1,mapTest$krigData,K$listZonePoint,K$zoneN,
       mapTest$krigSurfVoronoi,K$meanZone,pErr=0.9)
calCrit7(resD$matDistanceCorr,K$zoneNModif)
# not run
}
