% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCritMinMean}
\alias{calCritMinMean}
\title{calCritMinMean}
\usage{
calCritMinMean(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCritMinMean
}
\details{
description, a paragraph
}
\examples{
# not run
}
