% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{calculMatriceVario}
\alias{calculMatriceVario}
\title{calculMatriceVario}
\usage{
calculMatriceVario(distij, maxVario, fitVarioAlea)
}
\arguments{
\item{distij}{xxxx}

\item{maxVario}{xxxx}

\item{fitVarioAlea}{xxxx}
}
\value{
a ?
}
\description{
calculMatriceVario
}
\details{
description, a paragraph
}
\examples{
# not run
}
