% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calczf}
\alias{calczf}
\title{calczf}
\usage{
calczf(qProb, map, optiCrit, minSize, minSizeNG, disp = 0, pdf1 = NULL,
  FULL = FALSE, data = NULL)
}
\arguments{
\item{qProb}{xxxx}

\item{map}{xxxx}

\item{optiCrit}{xxxx}

\item{minSize}{xxxx}

\item{minSizeNG}{xxxx}

\item{disp}{xxxx}

\item{pdf1}{xxxx}

\item{FULL}{xxxx}

\item{data}{xxxx}
}
\value{
a list
}
\description{
calczf
}
\details{
description, a paragraph
}
\examples{
# not run
}
