% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{calDistZone1}
\alias{calDistZone1}
\title{calDistZone1}
\usage{
calDistZone1(indice1, indice2, listZonePoint = NULL, tabVal = NULL,
  surfaceVoronoi, meanZone2, pErr = 0.9)
}
\arguments{
\item{indice1}{xxxx}

\item{indice2}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfaceVoronoi}{xxxx}

\item{meanZone2}{xxxx}

\item{pErr}{xxxx}
}
\value{
a ?
}
\description{
calDistZone1
}
\details{
description, a paragraph
}
\examples{
# not run
}
