% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{calDistZone12}
\alias{calDistZone12}
\title{calDistZone12}
\usage{
calDistZone12(indice1, indice2, listZonePoint = NULL, tabVal = NULL,
  surfaceVoronoi, meanZone2)
}
\arguments{
\item{indice1}{xxxx}

\item{indice2}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfaceVoronoi}{xxxx}

\item{meanZone2}{xxxx}
}
\value{
a ?
}
\description{
calDistZone12
}
\details{
description, a paragraph
}
\examples{
# not run
}
