% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{calDistZone2}
\alias{calDistZone2}
\title{calDistZone2}
\usage{
calDistZone2(indice1, indice2, listZonePoint, tabVal, surfaceVoronoi,
  fitVarioAlea)
}
\arguments{
\item{indice1}{xxxx}

\item{indice2}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfaceVoronoi}{xxxx}

\item{fitVarioAlea}{xxxx}
}
\value{
a ?
}
\description{
calDistZone2
}
\details{
description, a paragraph
}
\examples{
# not run
}
