% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{calDistZoneMoranB}
\alias{calDistZoneMoranB}
\title{calDistZoneMoranB}
\usage{
calDistZoneMoranB(indice1, indice2, listZonePoint, tabVal, surfaceVoronoi,
  moyTot)
}
\arguments{
\item{indice1}{xxxx}

\item{indice2}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfaceVoronoi}{xxxx}

\item{moyTot}{xxxx}
}
\value{
a ?
}
\description{
calDistZoneMoranB
}
\details{
description, a paragraph
}
\examples{
# not run
}
