% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFrame.R
\name{calFrame}
\alias{calFrame}
\title{calFrame}
\usage{
calFrame(iZ, Z, zoneNModif, distIsoZ = 0.075)
}
\arguments{
\item{iZ}{index of zone for which the envelope is searched}

\item{Z}{zoning}

\item{zoneNModif}{modified zone neighborhood matrix (FALSE values on diagonal}

\item{distIsoZ}{threshold distance above which a zone is considered as isolated}
}
\value{
a apatial polygon corresponding to the frame within which grown zone must be contained
}
\description{
calFrame
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
Z=resZTest$zonePolygone
zN=resZTest$zoneNModif
f=calFrame(6,Z,zN)
plotZ(Z)
rgeos::plot(f,add=TRUE,col="red")
}
