% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calGearyGlo}
\alias{calGearyGlo}
\title{calGearyGlo}
\usage{
calGearyGlo(matN, vectMean, meanTot, vectSurface)
}
\arguments{
\item{matN}{xxxx}

\item{vectMean}{xxxx}

\item{meanTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
calGearyGlo
}
\details{
computes global Geary criterion
}
\examples{
# not run
}
