% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calGearyLoc}
\alias{calGearyLoc}
\title{vecteur d'indices de Geary locaux}
\usage{
calGearyLoc(matVoisin, vectMoy, moyTot, vectSurface)
}
\arguments{
\item{matVoisin}{xxxx}

\item{vectMoy}{xxxx}

\item{moyTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
vecteur d'indices de Geary locaux
}
\details{
description, a paragraph
}
\examples{
# not run
}
