% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calIndices.R
\name{calIndice}
\alias{calIndice}
\title{calIndice}
\usage{
calIndice(tabAlea, voisin, zoneNmodif, listZonePoint, listSurf, meanZone,
  surfVoronoi, matDistanceMoranB)
}
\arguments{
\item{tabAlea}{xxxx}

\item{voisin}{xxxx}

\item{zoneNmodif}{xxxx}

\item{listZonePoint}{xxxx}

\item{listSurf}{xxxx}

\item{meanZone}{xxxx}

\item{surfVoronoi}{xxxx}

\item{matDistanceMoranB}{xxxx}
}
\value{
a ?
}
\description{
calIndice
}
\details{
description, a paragraph
}
\examples{
# not run
}
