% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{calMatDistance}
\alias{calMatDistance}
\title{calMatDistance}
\usage{
calMatDistance(typedist = 1, zoneN = NULL, listZonePoint = NULL,
  tabVal = NULL, surfVoronoi = NULL, meanZone = NULL, pErr)
}
\arguments{
\item{typedist}{xxxx}

\item{zoneN}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfVoronoi}{xxxx}

\item{meanZone}{xxxx}

\item{pErr}{xxxx}
}
\value{
a ?
}
\description{
calMatDistance
}
\details{
description, a paragraph
}
\examples{
# not run
}
