% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranBLocal}
\alias{calMoranBLocal}
\title{fonction calant un vecteur d'indices de moran locaux(= par zone)}
\usage{
calMoranBLocal(voisinZone, matDistanceMoranB, vectSurface)
}
\arguments{
\item{voisinZone}{xxxx}

\item{matDistanceMoranB}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
fonction calant un vecteur d'indices de moran locaux(= par zone)
}
\details{
description, a paragraph
}
\examples{
# not run
}
