% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranBTot}
\alias{calMoranBTot}
\title{computes Moran criterion on whole zoning}
\usage{
calMoranBTot(NZone, matDistanceMoranB, vectSurface)
}
\arguments{
\item{NZone}{xxxx}

\item{matDistanceMoranB}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
computes Moran criterion on whole zoning
}
\details{
computes Moran criterion on zoning
}
\examples{
# not run
}
