% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranGlo}
\alias{calMoranGlo}
\title{computes specific Moran criterion}
\usage{
calMoranGlo(matNZone, vectMean, meanTot, vectSurface)
}
\arguments{
\item{matNZone}{xxxx}

\item{vectMean}{xxxx}

\item{meanTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
computes specific Moran criterion
}
\details{
description, a paragraph
}
\examples{
# not run
}
