% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranLoc}
\alias{calMoranLoc}
\title{calMoranLoc}
\usage{
calMoranLoc(matVoisin, vectMoy, moyTot, vectSurface)
}
\arguments{
\item{matVoisin}{xxxx}

\item{vectMoy}{xxxx}

\item{moyTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
calMoranLoc
}
\details{
description, a paragraph
}
\examples{
# not run
}
