% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNei.R
\name{calNei}
\alias{calNei}
\title{calNei}
\usage{
calNei(Z, data, surfVoronoi, ptN, simplitol, remove = TRUE, correct = FALSE)
}
\arguments{
\item{Z}{xxxx}

\item{data}{xxxx}

\item{surfVoronoi}{xxxx}

\item{ptN}{xxxx}

\item{simplitol}{xxxx}

\item{remove}{xxxx}

\item{correct}{xxxx}
}
\value{
a ?
}
\description{
calNei
}
\details{
description, a paragraph
}
\examples{
# not run
}
