% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calRapPolygone}
\alias{calRapPolygone}
\title{calRapPolygone}
\usage{
calRapPolygone(Surf, polyL)
}
\arguments{
\item{Surf}{xxxx}

\item{polyL}{xxxx}
}
\value{
a ?
}
\description{
calRapPolygone
}
\details{
description, a paragraph
}
\examples{
# not run
}
