% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRMmodel.R
\name{calRMmodel}
\alias{calRMmodel}
\title{transform VGM model into model usable by RandomFields}
\usage{
calRMmodel(vgmodel)
}
\arguments{
\item{vgmodel}{model provided by a call to vgm}
}
\value{
model suitable for RandomFields simulation
}
\description{
transform VGM model into model usable by RandomFields
}
\examples{
modv=gstat::vgm(model="Gau",range=100,psill=10,mean=7)
RMmodel=calRMmodel(modv)
}
