% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalNei.R
\name{calZoneN}
\alias{calZoneN}
\title{calZoneN}
\usage{
calZoneN(ptN, zoneN, listZonePoint)
}
\arguments{
\item{ptN}{pt neighborhood Logical matrix}

\item{zoneN}{empty zone neighborhood Logical matrix}

\item{listZonePoint}{list of indices of data points within zones}
}
\value{
a list with component zoneN holding filled zone neighborhood Logical matrix
}
\description{
calZoneN
}
\details{
calculate zone neighborhood
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
ptN=mapTest$krigN
nZ=length(K$zonePolygone)
zoneN=matrix(logical(nZ*nZ),nZ,nZ)
listZonePoint=K$listZonePoint
calZoneN(ptN,zoneN,listZonePoint)
}
