% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkContour}
\alias{checkContour}
\title{checkContour}
\usage{
checkContour(contourSp, step, refPoint, minSizeNG)
}
\arguments{
\item{contourSp}{xxxx}

\item{step}{xxxx}

\item{refPoint}{xxxx}

\item{minSizeNG}{xxxx}
}
\value{
a plot
}
\description{
checkContour
}
\details{
description, a paragraph
}
\examples{
# not run
}
