% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cleanSp}
\alias{cleanSp}
\title{cleanSp}
\usage{
cleanSp(sp, tol = 1e-05)
}
\arguments{
\item{sp}{SpatialPolygons}

\item{tol}{minimum area for removal}
}
\value{
a SpatialPolygons
}
\description{
cleanSp
}
\details{
removes from sp polygons that are too small (artefacts of gDifference)
}
\examples{
# not run
}
