% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeZoneInt.R
\name{computeZoneInt}
\alias{computeZoneInt}
\title{computeZoneInt}
\usage{
computeZoneInt(qi, Z0, i0, map)
}
\arguments{
\item{qi}{xxxx}

\item{Z0}{xxxx}

\item{i0}{xxxx}

\item{map}{xxxx}
}
\value{
a plot
}
\description{
computeZoneInt
}
\details{
description, a paragraph
}
\examples{
# not run
}
