% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contourArea}
\alias{contourArea}
\title{contourArea}
\usage{
contourArea(co)
}
\arguments{
\item{co}{contour line}
}
\value{
the area within the contour line
}
\description{
contourArea
}
\details{
area corresponding to closed contour line
}
\examples{
data(mapTest)
cL=list()
cL=contourAuto(cL,mapTest$step,mapTest$xsize,mapTest$ysize,mapTest$krigGrid,c(5,7),mapTest$boundary)
contourArea(cL[[8]])
}
