% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contourToSpp}
\alias{contourToSpp}
\title{contourToSpp}
\usage{
contourToSpp(co, step)
}
\arguments{
\item{co}{contour line (list with contour level and x,y coordinates}

\item{step}{grid resolution}
}
\value{
a list with components
\describe{
\item{sp}{SpatialPolygons corresponding to contour line}
\item{contour}{SpatialLines corresponding to contour line}
\item{polyBuff}{SpatialPolygons corresponding to buffer around contour line}
\item{surface}{SpatialPolygons area}
}
}
\description{
contourToSpp
}
\details{
transform contour line into SpatialPolygons
}
\examples{
data(mapTest)
cL=list()
cL=contourAuto(cL,mapTest$step,mapTest$xsize,mapTest$ysize,mapTest$krigGrid,c(5,7),mapTest$boundary)
contourToSpp(cL[[8]],0.1)
}
