% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBoundaryMap.R
\name{correctBoundaryMap}
\alias{correctBoundaryMap}
\title{correctBoundaryMap}
\usage{
correctBoundaryMap(Zi, map)
}
\arguments{
\item{Zi}{list of initiales zones}

\item{map}{object returned by function genMap}
}
\value{
new list of zones with correct boundary ang the parameter "width" used for correction
}
\description{
correctBoundaryMap
}
\details{
function for post treatment of zoning that fixes the problem linked to
 the border between two neighbour zones and between zones and the map boundary
}
\examples{
map=geozoning::mapTest
criti = correctionTree(qProb = c(0.5), map = map)
Z = criti$zk[[1]][[1]]$zonePolygone
lab = criti$zk[[1]][[1]]$lab
plotM(map = map, Z = Z, lab = lab, byLab = FALSE)
class(rgeos::gIntersection(Z[[1]],Z[[2]])) [1]
class(rgeos::gIntersection(Z[[1]],Z[[5]])) [1]
class(rgeos::gIntersection(Z[[2]],Z[[3]])) [1]
class(rgeos::gIntersection(Z[[2]],Z[[4]])) [1]
res = correctBoundaryMap(Zi = Z, map = map)
Z = res$Z
class(rgeos::gIntersection(Z[[1]],Z[[2]])) [1]
class(rgeos::gIntersection(Z[[1]],Z[[5]])) [1]
class(rgeos::gIntersection(Z[[2]],Z[[3]])) [1]
class(rgeos::gIntersection(Z[[2]],Z[[4]])) [1]
plotM(map = map, Z = Z, lab = lab, byLab = FALSE)
}
