% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionTree.R
\name{correctionTree}
\alias{correctionTree}
\title{correctionTree}
\usage{
correctionTree(qProb, map, pErr = 0.9, optiCrit = 2, minSize = 0.012,
  minSizeNG = 0.001, distIsoZ = 0.075, simplitol = 0.001, LEQ = 5,
  MAXP = 0.1, LASTPASS = TRUE, disp = 0, SAVE = TRUE, ONE = FALSE)
}
\arguments{
\item{qProb}{xxxx}

\item{map}{xxxx}

\item{pErr}{xxxx}

\item{optiCrit}{xxxx}

\item{minSize}{xxxx}

\item{minSizeNG}{xxxx}

\item{distIsoZ}{xxxx}

\item{simplitol}{xxxx}

\item{LEQ}{xxxx}

\item{MAXP}{xxxx}

\item{LASTPASS}{xxxx}

\item{disp}{xxxx}

\item{SAVE}{xxxx}

\item{ONE}{xxxx}
}
\value{
a ?
}
\description{
correctionTree
}
\details{
description, a paragraph
}
\examples{
# not run
}
