% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNei.R
\name{correctN}
\alias{correctN}
\title{correctN}
\usage{
correctN(Z, zoneN, dN = 0.001)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{zoneN}{zone neighborhood Logical matrix}

\item{dN}{maximum distance beyond which 2 zones cannot be considered as neighbors}
}
\value{
a new zone neighborhood Logical matrix
}
\description{
correctN
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
Z=resZTest$zonePolygone
H=correctN(Z,resZTest$zoneN,1e-8)
# not run
}
