% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calCost.R
\name{Cost_By_Mean}
\alias{Cost_By_Mean}
\title{Cost_By_Mean}
\usage{
Cost_By_Mean(map, Z, numZ)
}
\arguments{
\item{map}{object returned by function genMap}

\item{Z}{: an example of zoning (a list of zones)}

\item{numZ}{: number of the zone in which the cost will be computed}
}
\value{
the cost value as described
}
\description{
Cost_By_Mean
}
\details{
: function that returns the criterion COST by approximating the value at a point of the grid by
           the mean value of the zone.
}
\examples{
# cf. Transition_Zone_Near_Boundary() help page
}
