% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{crComment}
\alias{crComment}
\title{crComment}
\usage{
crComment(Z)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}
}
\value{
a zoning
}
\description{
crComment
}
\details{
create comment corresponding to holes in a zoning
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
Z1=crComment(Z)
# not run
}
