% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{createHoles}
\alias{createHoles}
\title{createHoles}
\usage{
createHoles(Z)
}
\arguments{
\item{Z}{xxxx}
}
\value{
a ?
}
\description{
createHoles
}
\details{
description, a paragraph
}
\examples{
# not run
}
