% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{datanorm}
\alias{datanorm}
\title{normalize data coords and border}
\usage{
datanorm(data, bd)
}
\arguments{
\item{data}{xxxx}

\item{bd}{xxxx}
}
\value{
a list
}
\description{
normalize data coords and border
}
\details{
description, a paragraph
}
\examples{
# not run
}
