% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{datanormXY}
\alias{datanormXY}
\title{normalize data coords}
\usage{
datanormXY(data)
}
\arguments{
\item{data}{xxxx}
}
\value{
a list
}
\description{
normalize data coords
}
\details{
description, a paragraph
}
\examples{
# not run
}
