% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataReg}
\alias{dataReg}
\title{A data frame with simulated data on a regular grid}
\format{a data frame containing a regular grid with 1936 rows and 3 variables
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{z}{numeric variable - simulated}
}}
\usage{
dataReg
}
\description{
A data frame with simulated data on a regular grid
}
\keyword{datasets}
