% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{denormD}
\alias{denormD}
\title{denormD}
\usage{
denormD(DataObj, boundary)
}
\arguments{
\item{DataObj}{xxxx}

\item{boundary}{xxxx}
}
\value{
a ?
}
\description{
denormD
}
\details{
description, a paragraph
}
\examples{
# not run
}
