% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{detectSmallZones}
\alias{detectSmallZones}
\title{detectSmallZones}
\usage{
detectSmallZones(zonePolygone, minSize)
}
\arguments{
\item{zonePolygone}{xxxx}

\item{minSize}{xxxx}
}
\value{
a ?
}
\description{
detectSmallZones
}
\details{
description, a paragraph
}
\examples{
# not run
}
