% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZoneCE.R
\name{detZoneClose}
\alias{detZoneClose}
\title{detZoneClose}
\usage{
detZoneClose(iC, Z, K)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}
}
\value{
a ?
}
\description{
detZoneClose
}
\details{
description, a paragraph
}
\examples{
# not run
}
