% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZoneCE.R
\name{detZoneEng2}
\alias{detZoneEng2}
\title{detZoneEng2}
\usage{
detZoneEng2(iC, Z, K)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}
}
\value{
a ?
}
\description{
detZoneEng2
}
\details{
description, a paragraph
}
\examples{
# not run
}
