% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZoneCE.R
\name{detZoneP0}
\alias{detZoneP0}
\title{detZoneP0}
\usage{
detZoneP0(iC, Z, K)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}
}
\value{
a ?
}
\description{
detZoneP0
}
\details{
description, a paragraph
}
\examples{
# not run
}
