% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZoneCE.R
\name{detZonePIso}
\alias{detZonePIso}
\title{detZonePIso}
\usage{
detZonePIso(iC, Z, K)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}
}
\value{
a ?
}
\description{
detZonePIso
}
\details{
description, a paragraph
}
\examples{
# not run
}
