% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCond.R
\name{dispResCond}
\alias{dispResCond}
\title{dispResCond}
\usage{
dispResCond(krigMatTest, matMoyCond, matEcartCond, matEcartNorm, step)
}
\arguments{
\item{krigMatTest}{xxxx}

\item{matMoyCond}{xxxx}

\item{matEcartCond}{xxxx}

\item{matEcartNorm}{xxxx}

\item{step}{xxxx}
}
\value{
a ?
}
\description{
dispResCond
}
\details{
description, a paragraph
}
\examples{
# not run
}
