% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{dispZ}
\alias{dispZ}
\title{dispZ}
\usage{
dispZ(step, matVal, nbLvl = 20, zonePolygone = NULL, K = NULL,
  coulBreaks = 0, texMain = "", boundary = NULL, id = TRUE,
  valeurQuant = NULL, palCoul = colorRampPalette(c("brown", "yellow")),
  indiceArg = 0, mu = 1)
}
\arguments{
\item{step}{xxxx}

\item{matVal}{xxxx}

\item{nbLvl}{xxxx}

\item{zonePolygone}{xxxx}

\item{K}{xxxx}

\item{coulBreaks}{xxxx}

\item{texMain}{xxxx}

\item{boundary}{xxxx}

\item{id}{xxxx}

\item{valeurQuant}{xxxx}

\item{palCoul}{xxxx}

\item{indiceArg}{xxxx}

\item{mu}{xxxx}
}
\value{
a ?
}
\description{
dispZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
