% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{dispZmap}
\alias{dispZmap}
\title{dispZmap}
\usage{
dispZmap(map, Z = NULL, qProb = NULL, valbp = NULL, scale = NULL,
  lev = 20, palCol = colorRampPalette(c("brown", "yellow")),
  legend.width = 1, parG = NULL, ptz = NULL)
}
\arguments{
\item{map}{map object returned by genMap function}

\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{qProb}{quantile associated probability vector}

\item{valbp}{values used for boxplots}

\item{scale}{field scale}

\item{lev}{number of color levels}

\item{palCol}{color palette}

\item{legend.width}{relative width of legend}

\item{parG}{graphics parameters (result of call to par)}

\item{ptz}{zone id location, if NULL automatically find the best locations}
}
\value{
an empty value
}
\description{
dispZmap
}
\details{
plots a color representation of values and zones
}
\examples{
seed=80
data(mapTest)
ZK=initialZoning(c(0.5,0.7),mapTest)
K=ZK$resZ
Z=K$zonePolygone
#order zone ids by attribute mean value
ord=order(K$meanZone)
  Z=orderZ(Z,ord)
  plotZ(Z,id=TRUE)
# not run
}
