% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{distanceExtension}
\alias{distanceExtension}
\title{distanceExtension}
\usage{
distanceExtension(matDistance, matDistanceNorm, matDistanceCorr)
}
\arguments{
\item{matDistance}{xxxx}

\item{matDistanceNorm}{xxxx}

\item{matDistanceCorr}{xxxx}
}
\value{
a ?
}
\description{
distanceExtension
}
\details{
description, a paragraph
}
\examples{
# not run
}
