% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{extensionLine}
\alias{extensionLine}
\title{extensionLine}
\usage{
extensionLine(contour = NULL, step = NULL, boundary, bspl)
}
\arguments{
\item{contour}{xxxx}

\item{step}{xxxx}

\item{boundary}{xxxx}

\item{bspl}{xxxx}
}
\value{
a ?
}
\description{
extensionLine
}
\details{
description, a paragraph
}
\examples{
# not run
}
