% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractionPoly}
\alias{extractionPoly}
\title{extractionPoly}
\usage{
extractionPoly(polyTot)
}
\arguments{
\item{polyTot}{xxxx}
}
\value{
a ?
}
\description{
extractionPoly
}
\details{
description, a paragraph
}
\examples{
# not run
}
