% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figCrit.R
\name{figCrit}
\alias{figCrit}
\title{figCrit}
\usage{
figCrit(seed = 89, gr = 1, m1 = NULL, m2 = NULL, m3 = NULL,
  m4 = NULL, NEW = FALSE, ONE = FALSE, title = NULL)
}
\arguments{
\item{seed}{xxxx}

\item{gr}{xxxx}

\item{m1}{xxxx}

\item{m2}{xxxx}

\item{m3}{xxxx}

\item{m4}{xxxx}

\item{NEW}{xxxx}

\item{ONE}{xxxx}

\item{title}{xxxx}
}
\value{
a plot
}
\description{
figCrit
}
\details{
description, a paragraph
}
\examples{
# not run
}
