% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figCritN.R
\name{figCritN}
\alias{figCritN}
\title{figCritN}
\usage{
figCritN(seed = 89, m1 = NULL, m2 = NULL, m3 = NULL, m4 = NULL,
  m5 = NULL, NEW = FALSE, ONE = FALSE,
  title = "Gaussian field simulation", pdf = NULL)
}
\arguments{
\item{seed}{xxxx}

\item{m1}{xxxx}

\item{m2}{xxxx}

\item{m3}{xxxx}

\item{m4}{xxxx}

\item{m5}{xxxx}

\item{NEW}{xxxx}

\item{ONE}{xxxx}

\item{title}{xxxx}

\item{pdf}{xxxx}
}
\value{
a plot
}
\description{
figCritN
}
\details{
description, a paragraph
}
\examples{
# not run
}
