% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{findCinZ}
\alias{findCinZ}
\title{find contour for a given vRef quantile value}
\usage{
findCinZ(iC, Z, K, map, vRef, envel)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}

\item{map}{xxxx}

\item{vRef}{xxxx}

\item{envel}{xxxx}
}
\value{
a ?
}
\description{
find contour for a given vRef quantile value
}
\details{
description, a paragraph
}
\examples{
# not run
}
