% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{findNptInZone}
\alias{findNptInZone}
\title{findNptInZone}
\usage{
findNptInZone(Z, K, i1, i2, map)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{i1}{xxxx}

\item{i2}{xxxx}

\item{map}{xxxx}
}
\value{
a ?
}
\description{
findNptInZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
