% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{findNumZ}
\alias{findNumZ}
\title{findNumZ}
\usage{
findNumZ(Z, id)
}
\arguments{
\item{Z}{xxxx}

\item{id}{xxxx}
}
\value{
a ?
}
\description{
findNumZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
