% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{findZCenterpt}
\alias{findZCenterpt}
\title{findZCenterpt}
\usage{
findZCenterpt(data, K, num = NULL)
}
\arguments{
\item{data}{SpatialPointsDataFrame}

\item{K}{zoning object, as returned by the calNei function}

\item{num}{zone number or NULL for all zones}
}
\value{
a matrix of x and y coordinates for chosen points with as many rows as zones
}
\description{
findZCenterpt
}
\details{
find point within zone for pretty labelling
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.7),mapTest,SAVE=TRUE) 
K=criti$zk[[2]][[1]]
data=mapTest$krigData
geozoning:::findZCenterpt(data,K)
# not run
}
\keyword{internal}
