% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRandKmap.R
\name{genEmptyGrid}
\alias{genEmptyGrid}
\title{generate grid from raw data}
\usage{
genEmptyGrid(step, xsize, ysize)
}
\arguments{
\item{step}{numeric step for grid}

\item{xsize}{numeric value giving the data range on the x axis}

\item{ysize}{numeric value giving the data range on the y axis}
}
\value{
a list that contains x and y kriged positions based on original ones,#' plus nx and ny (number of x and y positions).
}
\description{
generate grid from raw data
}
\examples{
genEmptyGrid(calStep(1000,1,1),1,1)
}
