% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMap.R
\name{genMap}
\alias{genMap}
\title{wrapper for randKmap, generate 2D map}
\usage{
genMap(DataObj = NULL, seed = 80, krig = 1, Vpsill = 5, Vrange = 0.2,
  Vnugget = 0, Vmean = 8, disp = 0, FULL = FALSE)
}
\arguments{
\item{DataObj}{=NULL: simulated data with seed or = a data frame with real data}

\item{seed}{numeric,}

\item{krig}{numeric, 1: kriging with vgm model, 2: inverse distance kriging}

\item{Vpsill}{numeric,}

\item{Vrange}{numeric,}

\item{Vnugget}{numeric,}

\item{Vmean}{numeric,}

\item{disp}{numeric,}

\item{FULL}{logical, if TRUE the returned list is complete}
}
\value{
a map in a list
\describe{
\item{tabAlea}{raw data, SpatialPointsDataFrame}
\item{surfaceVoronoi}{Voronoi polygon surfaces}
\item{krigTabAlea}{kriged data, SpatialPointsDataFrame}
\item{fitVarioAlea}{variogram}
\item{DataObj}{DataObj}
}
}
\description{
wrapper for randKmap, generate 2D map
}
\details{
wrapper for randKmap, generate 2D map
}
\examples{
# not run
}
