% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMapR.R
\name{genMapR}
\alias{genMapR}
\title{genMapR}
\usage{
genMapR(DataObj, seed = NULL, disp = 0, krig = 1, boundary = boundary,
  nPointsK = 7000, FULL = FALSE)
}
\arguments{
\item{DataObj}{xxxx}

\item{seed}{xxxx}

\item{disp}{xxxx}

\item{krig}{xxxx}

\item{boundary}{xxxx}

\item{nPointsK}{xxxx}

\item{FULL}{xxxx}
}
\value{
a ?
}
\description{
genMapR
}
\details{
description, a paragraph
}
\examples{
# not run
}
