% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMapRNoK.R
\name{genMapRNoK}
\alias{genMapRNoK}
\title{genMapRNoK}
\usage{
genMapRNoK(DataObj, seed, boundary, disp = 0, krig = 1, FULL = FALSE)
}
\arguments{
\item{DataObj}{xxxx}

\item{seed}{xxxx}

\item{boundary}{xxxx}

\item{disp}{xxxx}

\item{krig}{xxxx}

\item{FULL}{xxxx}
}
\value{
a ?
}
\description{
genMapRNoK
}
\details{
description, a paragraph
}
\examples{
# not run
}
