% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{genQseq}
\alias{genQseq}
\title{genQseq}
\usage{
genQseq(qProb, K, map, i1, i2, LEQ = 5, MAXP = 0.1, disp = 0)
}
\arguments{
\item{qProb}{probability vector used to generate quantile values}

\item{K}{zoning object, as returned by the calNei function}

\item{map}{object returned by function genMap}

\item{i1}{current zone index}

\item{i2}{englobing zone index}

\item{LEQ}{length of quantile sequence}

\item{MAXP}{maximum shift from center for quantile sequence}

\item{disp}{0: no info, 1: some info}
}
\value{
a plot
}
\description{
genQseq
}
\details{
description, a paragraph
}
\examples{
data(mapTest)
qProb=c(0.4,0.7)
ZK=initialZoning(qProb,mapTest)
K=ZK$resZ
print(K$lab)
genQseq(qProb,K,mapTest,1,2) # from label 3 to label 2
# not run
}
