% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{genQseq0}
\alias{genQseq0}
\title{genQseq0}
\usage{
genQseq0(qProb, K, map, i1, i2, leq, disp = 0)
}
\arguments{
\item{qProb}{xxxx}

\item{K}{xxxx}

\item{map}{xxxx}

\item{i1}{xxxx}

\item{i2}{xxxx}

\item{leq}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
genQseq0
}
\details{
description, a paragraph
}
\examples{
# not run
}
