% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getBestMloop}
\alias{getBestMloop}
\title{getBestMloop}
\usage{
getBestMloop(seed, thr = 0.5)
}
\arguments{
\item{seed}{xxxx}

\item{thr}{xxxx}
}
\value{
a ?
}
\description{
getBestMloop
}
\details{
description, a paragraph
}
\examples{
# not run
}
