% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getClosePt}
\alias{getClosePt}
\title{getClosePt}
\usage{
getClosePt(Z, iC, iZC, disp = FALSE)
}
\arguments{
\item{Z}{zoning (list of SpatialPolygons)}

\item{iC}{current zone indes}

\item{iZC}{close zone index}

\item{disp}{information level (FALSE-no info)}
}
\value{
a SpatialPoints of length 1
}
\description{
getClosePt
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
getClosePt(Z,1,3)
plotZ(Z)
points( getClosePt(Z,1,3),col="blue",pch=20)
# not run
}
