% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getClosestZone}
\alias{getClosestZone}
\title{getClosestZone}
\usage{
getClosestZone(iC, Z, K)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}
}
\value{
a ?
}
\description{
getClosestZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
