% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getCoords}
\alias{getCoords}
\title{getCoords}
\usage{
getCoords(sp, k = 1)
}
\arguments{
\item{sp}{SpatialPolygons}

\item{k}{polygon number}
}
\value{
some coordinates
}
\description{
getCoords
}
\details{
get SpatialPolygons coordinates
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
geozoning:::getCoords(Z[[1]])
}
\keyword{internal}
