% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getId}
\alias{getId}
\title{getId}
\usage{
getId(Z, iZ)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iZ}{zone number}
}
\value{
a character vector giving the zone identifier
}
\description{
getId
}
\details{
get zone identifier in a zoning
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.5),mapTest,SAVE=TRUE)
Z=criti$zk[[2]][[1]]$zonePolygon
getId(Z,6)
# not run
}
