% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getIds}
\alias{getIds}
\title{getIds}
\usage{
getIds(Z, nums = NULL)
}
\arguments{
\item{Z}{xxxx}

\item{nums}{xxxx}
}
\value{
a ?
}
\description{
getIds
}
\details{
description, a paragraph
}
\examples{
# not run
}
