% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getKMloop}
\alias{getKMloop}
\title{getKMloop}
\usage{
getKMloop(seed, m1 = NULL, m2 = NULL, m3 = NULL, m4 = NULL, m5 = NULL,
  k = 5, pE = 0.9)
}
\arguments{
\item{seed}{xxxx}

\item{m1}{xxxx}

\item{m2}{xxxx}

\item{m3}{xxxx}

\item{m4}{xxxx}

\item{m5}{xxxx}

\item{k}{xxxx}

\item{pE}{xxxx}
}
\value{
a ?
}
\description{
getKMloop
}
\details{
description, a paragraph
}
\examples{
# not run
}
