% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getNq}
\alias{getNq}
\title{getNq}
\usage{
getNq(critList)
}
\arguments{
\item{critList}{component critList of result from correctionTree}
}
\value{
a vector with the size of quantile vectors for each zoning corresponding to critList
}
\description{
getNq
}
\details{
determine size of quantile  in result from correctionTree
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.7),mapTest,SAVE=FALSE) 
getNq(criti$critList)
# not run
}
