% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNs.R, R/utilZ.R
\name{getNs}
\alias{getNs}
\alias{getNs}
\title{getNs}
\usage{
getNs(zoneN, iZ)

getNs(zoneN, iZ)
}
\arguments{
\item{zoneN}{zone neighborhood Logical matrix}

\item{iZ}{index of current zone in zoning}

\item{resZ}{xxxx}

\item{iC}{xxxx}
}
\value{
a ?

a Logical vector of current zone neighbors
}
\description{
getNs

getNs
}
\details{
description, a paragraph

get zone numbers of neighbors of a given zone
}
\examples{
# not run
data(mapTest)
K=resZTest
Ns=getNs(K$zoneNModif,5) # find neighbors of zone 5
}
