% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getNumZone}
\alias{getNumZone}
\title{getNumZone}
\usage{
getNumZone(ptsp, Z)
}
\arguments{
\item{ptsp}{xxxx}

\item{Z}{xxxx}
}
\value{
a ?
}
\description{
getNumZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
