% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPoly}
\alias{getPoly}
\title{getPoly}
\usage{
getPoly(Z, iZ, k)
}
\arguments{
\item{Z}{xxxx}

\item{iZ}{xxxx}

\item{k}{xxxx}
}
\value{
a numeric?
}
\description{
getPoly
}
\details{
description, a paragraph
}
\examples{
# not run
}
