% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPolySp}
\alias{getPolySp}
\title{getPolySp}
\usage{
getPolySp(sp, k = 1)
}
\arguments{
\item{sp}{xxxx}

\item{k}{xxxx}
}
\value{
a numeric?
}
\description{
getPolySp
}
\details{
description, a paragraph
}
\examples{
# not run
}
