% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getPtsZone}
\alias{getPtsZone}
\title{getPtsZone}
\usage{
getPtsZone(ptsp, zone)
}
\arguments{
\item{ptsp}{xxxx}

\item{zone}{xxxx,}
}
\value{
a map in a list
\describe{
\item{pts}{ptsub}
\item{mask}{mask}
}
}
\description{
getPtsZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
