% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getSurf}
\alias{getSurf}
\title{getSurf}
\usage{
getSurf(Z, iZ)
}
\arguments{
\item{Z}{xxxx}

\item{iZ}{xxxx}
}
\value{
a ?
}
\description{
getSurf
}
\details{
description, a paragraph
}
\examples{
# not run
}
