% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getZoneId}
\alias{getZoneId}
\title{getZoneId}
\usage{
getZoneId(zone)
}
\arguments{
\item{zone}{SpatialPolygons}
}
\value{
the zone identifier ( a character vector of length 1)
}
\description{
getZoneId
}
\details{
get the zone unique identifier
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.5),mapTest,SAVE=TRUE)
Z=criti$zk[[2]][[1]]$zonePolygon
getZoneId(Z[[4]])
# not run
}
