% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getZonePts}
\alias{getZonePts}
\title{getZonePts}
\usage{
getZonePts(ptsp, zone)
}
\arguments{
\item{ptsp}{SpatialPointsDataFrame with data values}

\item{zone}{SpatialPolygons defining a zone}
}
\value{
a list with components
\describe{
\item{pts}{SpatialPointsDataFrame with the data points within the zone}
\item{mask}{Logical vector of the within zone data points indices}
}
}
\description{
getZonePts
}
\details{
get all data points within a zone
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
data(mapTest)
ptsp=mapTest$krigData
res=getZonePts(ptsp,Z[[5]])
plotZ(Z)
points(res$pts,col="blue",pch=20)
# not run
}
