% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getZsize}
\alias{getZsize}
\title{getZsize}
\usage{
getZsize(Z)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}
}
\value{
a vector with x and y maximum values
}
\description{
getZsize
}
\details{
compute maximum x and y values of zoning Z
}
\examples{
# not run
#data(resZTest)
#Z=resZTest$zonePolygone
#geozoning:::getZsize(Z)
}
\keyword{internal}
