% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridXY.R
\name{gridXY}
\alias{gridXY}
\title{generate empty grid}
\usage{
gridXY(mat)
}
\arguments{
\item{mat}{matrix with x and y coordinates in the first two columns, data in third column}
}
\value{
a grid
}
\description{
generate empty grid
}
\details{
generate rectangular empty grid corresponding to x and y values in matrix
}
\examples{
data(dataReg)
geozoning:::gridXY(dataReg)
}
\keyword{internal}
