% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{holeSp}
\alias{holeSp}
\title{holeSp}
\usage{
holeSp(sp)
}
\arguments{
\item{sp}{SpatialPolygons}
}
\value{
the number of holes within sp
}
\description{
holeSp
}
\details{
number of holes in SpatialPolygons
}
\examples{
ZK=initialZoning(qProb=c(0.4,0.2,0.7),mapTest)
Z=ZK$resZ$zonePolygone
holeSp(Z[[5]]) #zone 5 has 1 hole
# not run
}
