% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{Identify}
\alias{Identify}
\title{Identify}
\usage{
Identify(id, Z)
}
\arguments{
\item{id}{xxxx}

\item{Z}{xxxx}
}
\value{
a ?
}
\description{
Identify
}
\details{
description, a paragraph
}
\examples{
# not run
}
