% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialZoning.R
\name{initialZoning}
\alias{initialZoning}
\title{initialZoning}
\usage{
initialZoning(qProb, map, pErr = 0.9, simplitol = 0.001, optiCrit = 2,
  disp = 0, GridData = F)
}
\arguments{
\item{qProb}{xxxx}

\item{map}{xxxx}

\item{pErr}{xxxx}

\item{simplitol}{xxxx}

\item{optiCrit}{xxxx}

\item{disp}{xxxx}

\item{GridData}{xxxx}
}
\value{
a ?
}
\description{
initialZoning
}
\details{
description, a paragraph
}
\examples{
# not run
}
