% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interContourZ}
\alias{interContourZ}
\title{interContourZ}
\usage{
interContourZ(Z, contour)
}
\arguments{
\item{Z}{xxxx}

\item{contour}{xxxx}
}
\value{
a ?
}
\description{
interContourZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
