% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{interZoneC}
\alias{interZoneC}
\title{interZoneC}
\usage{
interZoneC(Z, iC, iZC, closePt)
}
\arguments{
\item{Z}{xxxx}

\item{iC}{xxxx}

\item{iZC}{xxxx}

\item{closePt}{xxxx}
}
\value{
a ?
}
\description{
interZoneC
}
\details{
description, a paragraph
}
\examples{
# not run
}
