% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCond.R
\name{krigeBord}
\alias{krigeBord}
\title{krigeBord}
\usage{
krigeBord(moyCondTab, step, krigTabAlea)
}
\arguments{
\item{moyCondTab}{xxxx}

\item{step}{xxxx}

\item{krigTabAlea}{xxxx}
}
\value{
a ?
}
\description{
krigeBord
}
\details{
description, a paragraph
}
\examples{
# not run
}
