% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNei.R
\name{labZone}
\alias{labZone}
\title{labZone}
\usage{
labZone(listK, qProb, matVal)
}
\arguments{
\item{listK}{xxxx}

\item{qProb}{xxxx}

\item{matVal}{xxxx}
}
\value{
a ?
}
\description{
labZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
