% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNei.R
\name{labZone0}
\alias{labZone0}
\title{labZone0}
\usage{
labZone0(K, qProb, dataF)
}
\arguments{
\item{K}{zoning object, as returned by the calNei function}

\item{qProb}{probability vector used to generate quantile values for Z}

\item{dataF}{data used to generate labels and zoning}
}
\value{
a zoning object with labelled zones in lab component
}
\description{
labZone0
}
\details{
assigns a class label (integer) to a zone depending on the zone mean value
and on the quantile values. Default label is 1, corresponding to mean value samller #' or equal to first quantile. For k ordered quantile values, if mean value is greater #' than quantile k plus 10% of the data range, zone label is k.
}
\examples{
data(mapTest)
dataF=mapTest$krigGrid
data(resZTest)
K=resZTest
p = K$qProb
labZone(K,p,dataF)
# not run
}
