% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastPass.R
\name{lastPass}
\alias{lastPass}
\title{lastPass}
\usage{
lastPass(carte, qProb, listOfZ, crit, cost, costL, nz, mdist, pErr, optiCrit,
  simplitol, disp = F)
}
\arguments{
\item{carte}{xxxx}

\item{qProb}{xxxx}

\item{listOfZ}{xxxx}

\item{crit}{xxxx}

\item{cost}{xxxx}

\item{costL}{xxxx}

\item{nz}{xxxx}

\item{mdist}{xxxx}

\item{pErr}{xxxx}

\item{optiCrit}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
lastPass
}
\details{
description, a paragraph
}
\examples{
# not run
}
