% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{linesC}
\alias{linesC}
\title{linesC}
\usage{
linesC(listeContour, col = "blue")
}
\arguments{
\item{listeContour}{xxxx}

\item{col}{xxxx}
}
\value{
a ?
}
\description{
linesC
}
\details{
description, a paragraph
}
\examples{
# not run
}
