% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{linesSp}
\alias{linesSp}
\title{linesSp}
\usage{
linesSp(sp, k = 1, lty = 1, col = "red", lwd = 1)
}
\arguments{
\item{sp}{SpatialPolygons object}

\item{k}{polygon number}

\item{lty}{line type}

\item{col}{color}

\item{lwd}{line width}
}
\description{
linesSp
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
plotZ(Z)
linesSp(Z[[4]])
# not run
}
