% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lineToSp}
\alias{lineToSp}
\title{lineToSp}
\usage{
lineToSp(lin)
}
\arguments{
\item{lin}{list with x and y line coordinates}
}
\value{
a SpatialPolygons
}
\description{
lineToSp
}
\details{
transform closed line into SpatialPolygons
}
\examples{
data(mapTest)
cL=list()
cL=contourAuto(cL,mapTest$step,mapTest$xsize,mapTest$ysize,mapTest$krigGrid,c(5,7),mapTest$boundary)
lin=data.frame(x=cL[[8]]$x,y=cL[[8]]$y)
sp=geozoning:::lineToSp(lin)
}
\keyword{internal}
