% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{listContourArea}
\alias{listContourArea}
\title{listContourArea}
\usage{
listContourArea(cL)
}
\arguments{
\item{cL}{list of contour lines}
}
\value{
a list of areas
}
\description{
listContourArea
}
\details{
area of all contour lines in list
}
\examples{
data(mapTest)
cL=list()
cL=contourAuto(cL,mapTest$step,mapTest$xsize,mapTest$ysize,mapTest$krigGrid,c(5,7),mapTest$boundary)
geozoning:::listContourArea(cL)
# not run
}
\keyword{internal}
