% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopQ.R
\name{loopQ1}
\alias{loopQ1}
\title{loopQ1}
\usage{
loopQ1(map, disp = 1, step = 0.075, QUIET = FALSE)
}
\arguments{
\item{map}{xxxx}

\item{disp}{xxxx}

\item{step}{xxxx}

\item{QUIET}{xxxx}
}
\value{
a ?
}
\description{
loopQ1
}
\details{
description, a paragraph
}
\examples{
# not run
}
