% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopQ.R
\name{loopQ5}
\alias{loopQ5}
\title{loopQ5}
\usage{
loopQ5(map, disp = 1, step = 0.075, SAVE = F, QUIET = F)
}
\arguments{
\item{map}{xxxx}

\item{disp}{xxxx}

\item{step}{xxxx}

\item{SAVE}{xxxx}

\item{QUIET}{xxxx}
}
\value{
a ?
}
\description{
loopQ5
}
\details{
description, a paragraph
}
\examples{
# not run
}
