% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{maxDistSP}
\alias{maxDistSP}
\title{maxDistSP}
\usage{
maxDistSP(sp)
}
\arguments{
\item{sp}{SpatialPolygons}
}
\value{
the maximum distance within sp
}
\description{
maxDistSP
}
\details{
maximum distance within kth polygon of current zone
}
\examples{
ZK=initialZoning(qProb=c(0.4,0.2,0.7),mapTest)
Z=ZK$resZ$zonePolygone
maxDistSP(Z[[5]])
# not run
}
\keyword{internal}
