% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{maxDistZone}
\alias{maxDistZone}
\title{maxDistZone}
\usage{
maxDistZone(Z, iZ, k)
}
\arguments{
\item{Z}{xxxx}

\item{iZ}{xxxx}

\item{k}{xxxx}
}
\value{
a numeric?
}
\description{
maxDistZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
