% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{meanL}
\alias{meanL}
\title{meanL}
\usage{
meanL(zlab, listZonePoint, tabVal, surfVoronoi)
}
\arguments{
\item{zlab}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfVoronoi}{xxxx}
}
\value{
a ?
}
\description{
meanL
}
\details{
description, a paragraph
}
\examples{
# not run
}
