% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{meansdSimu}
\alias{meansdSimu}
\title{meansdSimu}
\usage{
meansdSimu(vseed = NULL, krig = 1)
}
\arguments{
\item{vseed}{xxxx}

\item{krig}{xxxx}
}
\value{
a ?
}
\description{
meansdSimu
}
\details{
description, a paragraph
}
\examples{
# not run
}
