% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{meanvarSimu}
\alias{meanvarSimu}
\title{meanVarsimu}
\usage{
meanvarSimu(map)
}
\arguments{
\item{map}{object generated by genMap}
}
\value{
a vector with  4 elements, the first two  give mean and standard deviation of generated raw data, the last two  give mean and standard deviation of kriged data
}
\description{
meanVarsimu
}
\details{
computes mean and standard deviation of a set of map simulations
}
\examples{
meansdSimu(c(1,2))
# not run
}
