% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{MeanVarWPts}
\alias{MeanVarWPts}
\title{MeanVarWPts}
\usage{
MeanVarWPts(map, zone, w = NULL)
}
\arguments{
\item{map}{object returned by function genMap}

\item{zone}{SpatialPolygons defining a zone}

\item{w}{weighting vector (default NULL)}
}
\value{
a list with components
\describe{
\item{mean}{(weighted) mean of the within zone attribute value}
\item{var}{(weighted) variance of the within zone attribute value}
}
}
\description{
MeanVarWPts
}
\details{
computes (weighted) mean and variance of zone data
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
data(mapTest)
MeanVarWPts(mapTest,Z[[1]])
# Weights are areas of the Voronoi polygons corresponding to data points
MeanVarWPts(mapTest,Z[[1]],mapTest$krigSurfVoronoi) #slightly different result
}
