% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{modlm}
\alias{modlm}
\title{modlm}
\usage{
modlm(ptsp, Z)
}
\arguments{
\item{ptsp}{xxxx}

\item{Z}{a zone?}
}
\value{
a lm object
}
\description{
modlm
}
\details{
description, a paragraph
}
\examples{
# not run
}
