% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{moveHoles}
\alias{moveHoles}
\title{moveHoles}
\usage{
moveHoles(zoneMain, zoneSuppr)
}
\arguments{
\item{zoneMain}{SpatialPolygons}

\item{zoneSuppr}{SpatialPolygons  inside main}
}
\value{
a new SpatialPolygons object
}
\description{
moveHoles
}
\details{
creates SpatialPolygons excluding holes
}
\examples{
# not run
}
\keyword{internal}
