% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normD}
\alias{normD}
\title{normD}
\usage{
normD(DataObj, boundary)
}
\arguments{
\item{DataObj}{xxxx}

\item{boundary}{xxxx}
}
\value{
a ?
}
\description{
normD
}
\details{
description, a paragraph
}
\examples{
# not run
}
