% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normDistMat}
\alias{normDistMat}
\title{normDistMat}
\usage{
normDistMat(matDistanceCorr, optiCrit)
}
\arguments{
\item{matDistanceCorr}{xxxx}

\item{optiCrit}{xxxx}
}
\value{
a ?
}
\description{
normDistMat
}
\details{
description, a paragraph
}
\examples{
# not run
}
