% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normZcoords}
\alias{normZcoords}
\title{normZcoords}
\usage{
normZcoords(Z, boundary)
}
\arguments{
\item{Z}{xxxx}

\item{boundary}{xxxx}
}
\value{
a list
}
\description{
normZcoords
}
\details{
description, a paragraph
}
\examples{
# not run
}
