% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nPolySp}
\alias{nPolySp}
\title{nPolySp}
\usage{
nPolySp(sp)
}
\arguments{
\item{sp}{SpatialPolygons}
}
\value{
the number of polygons within the current zone
}
\description{
nPolySp
}
\details{
number of polygons in SpatialPolygons

# not run
}
\examples{
ZK=initialZoning(qProb=c(0.4,0.2,0.7),mapTest)
Z=ZK$resZ$zonePolygone
print(paste(nPolySp(Z[[2]]),"polygons"))
}
