% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nPolyZone}
\alias{nPolyZone}
\title{nPolyZone}
\usage{
nPolyZone(Z, ind)
}
\arguments{
\item{Z}{xxxx}

\item{ind}{xxxx}
}
\value{
a numeric?
}
\description{
nPolyZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
