% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{obrien}
\alias{obrien}
\title{obrien}
\usage{
obrien(pts, ply)
}
\arguments{
\item{pts}{xxxx}

\item{ply}{xxxx}
}
\value{
a ?
}
\description{
obrien
}
\details{
description, a paragraph
}
\examples{
# not run
}
