% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{obrien2}
\alias{obrien2}
\title{obrien2}
\usage{
obrien2(pts, ply)
}
\arguments{
\item{pts}{xxxx}

\item{ply}{xxxx}
}
\value{
a ?
}
\description{
obrien2
}
\details{
description, a paragraph
}
\examples{
# not run
}
