% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optCost.R
\name{optCost}
\alias{optCost}
\title{optCost}
\usage{
optCost(seed, map, Cmin = 0, f = 1, disp = 1)
}
\arguments{
\item{seed}{xxxx}

\item{map}{xxxx}

\item{Cmin}{xxxx}

\item{f}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
optCost
}
\details{
description, a paragraph
}
\examples{
# not run
}
