% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optCostL.R
\name{optCostL}
\alias{optCostL}
\title{optCostL}
\usage{
optCostL(seed, map, Cmin = 0, f = 2)
}
\arguments{
\item{seed}{xxxx}

\item{map}{xxxx}

\item{Cmin}{xxxx}

\item{f}{xxxx}
}
\value{
a ?
}
\description{
optCostL
}
\details{
description, a paragraph
}
\examples{
# not run
}
