% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiGrow.R
\name{optiGrow}
\alias{optiGrow}
\title{optiGrow}
\usage{
optiGrow(Z, K, iC, qProb, refPoint, map, optiCrit, minSize = 0.012,
  minSizeNG = 0.001, distIsoZ = 0.075, LEQ = 5, MAXP = 0.1,
  simplitol = 1e-12, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{iC}{xxxx}

\item{qProb}{xxxx}

\item{refPoint}{xxxx}

\item{map}{xxxx}

\item{optiCrit}{xxxx}

\item{minSize}{xxxx}

\item{minSizeNG}{xxxx}

\item{distIsoZ}{xxxx}

\item{LEQ}{xxxx}

\item{MAXP}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
optiGrow
}
\details{
description, a paragraph
}
\examples{
# not run
}
