% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optRank.R
\name{optRank}
\alias{optRank}
\title{optCrit}
\usage{
optRank(vseed = NULL, k = 5)
}
\arguments{
\item{vseed}{xxxx}

\item{k}{xxxx}
}
\value{
a ?
}
\description{
optCrit
}
\details{
description, a paragraph
}
\examples{
# not run
}
