% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{orderZ}
\alias{orderZ}
\title{orderZ}
\usage{
orderZ(Z, ord)
}
\arguments{
\item{Z}{zoning geometry}

\item{ord}{sorting order}
}
\value{
a zoning geometry (list of SpatialPolygons)
}
\description{
orderZ
}
\details{
sorts zones according to ord vector
}
\examples{
# not run, take a while...
#map=genMap(DataObj=NULL,seed=40,disp=FALSE,krig=1,Vnugget=1.2,typeMod="Gau")
#qProb=c(0.275,0.8)
#criti=correctionTree(qProb,map,LASTPASS=FALSE)
#res=searchNODcrit1(qProb,criti)
#b=res$ind[[1]][1]
#K=criti$zk[[2]][[b]]
#Z=K$zonePolygone
#plotZ(Z)
#ord=valZ(map,K)$ord
#Z=orderZ(Z,ord)
#plotZ(Z)
}
