% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCrit.R
\name{plotCrit}
\alias{plotCrit}
\title{plotCrit}
\usage{
plotCrit(m1 = NULL, m2 = NULL, m3 = NULL, m4 = NULL, m5 = NULL,
  ONE = FALSE, title = "Gaussian field simulation")
}
\arguments{
\item{m1}{dataset with loopQ1 results}

\item{m2}{dataset with loopQ2 results}

\item{m3}{dataset with loopQ3 results}

\item{m4}{dataset with loopQ4 results}

\item{m5}{dataset with loopQ5 results}

\item{ONE}{single plot}

\item{title}{plot title}
}
\value{
a vector of probabilities corresponding to best results
}
\description{
plotCrit
}
\details{
reads loopQ1-5 results, filters results by keeping th best criteria ) and plots them together with corresponding costs.
}
\examples{
# not run

}
