% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{plotListC}
\alias{plotListC}
\title{plotListC}
\usage{
plotListC(cL, col = "red")
}
\arguments{
\item{cL}{list of contour lines}

\item{col}{color to use}
}
\value{
a plot
}
\description{
plotListC
}
\details{
add contour lines to a plot
}
\examples{
data(mapTest)
cL=list()
cL=contourAuto(cL,mapTest$step,mapTest$xsize,mapTest$ysize,mapTest$krigGrid,c(5,7),mapTest$boundary)
plot(mapTest$boundary,type="l")
plotListC(cL)
# not run
}
