% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotListeC}
\alias{plotListeC}
\title{plotListeC}
\usage{
plotListeC(listeC)
}
\arguments{
\item{listeC}{xxxx}
}
\value{
a plot
}
\description{
plotListeC
}
\details{
description, a paragraph
}
\examples{
# not run
}
