% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{plot a map}
\usage{
plotMap(map)
}
\arguments{
\item{map}{a map object, such as returned by genMap}
}
\value{
a plot
}
\description{
plot a map
}
\details{
plot 3 different graphics of a map object
}
\examples{
m=genMap(seed=1,krig=2,disp=0)
plotMap(m)
# not run

}
