% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotmat}
\alias{plotmat}
\title{plotmat}
\usage{
plotmat(m, pdf)
}
\arguments{
\item{m}{xxxx}

\item{pdf}{xxxx}
}
\value{
a ?
}
\description{
plotmat
}
\details{
description, a paragraph
}
\examples{
# not run
}
