% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotmdist}
\alias{plotmdist}
\title{plotmdist}
\usage{
plotmdist(md, pdf)
}
\arguments{
\item{md}{xxxx}

\item{pdf}{xxxx}
}
\value{
a ?
}
\description{
plotmdist
}
\details{
description, a paragraph
}
\examples{
# not run
}
