% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotOpt}
\alias{plotOpt}
\title{plotOpt}
\usage{
plotOpt(seed, pdf, k = 5)
}
\arguments{
\item{seed}{xxxx}

\item{pdf}{xxxx}

\item{k}{xxxx}
}
\value{
a ?
}
\description{
plotOpt
}
\details{
description, a paragraph
}
\examples{
# not run
}
