% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{plotSp}
\alias{plotSp}
\title{plotSp}
\usage{
plotSp(sp, k = 1, xlim, ylim)
}
\arguments{
\item{sp}{SpatialPolygons object}

\item{k}{polygon number}

\item{xlim}{x range}

\item{ylim}{y range}
}
\description{
plotSp
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
plotSp(Z[[1]],xlim=c(0,1),ylim=c(0,1))
}
