% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotVario}
\alias{plotVario}
\title{plotVario}
\usage{
plotVario(map, ylim = NULL)
}
\arguments{
\item{map}{xxxx}

\item{ylim}{xxxx}
}
\value{
a ?
}
\description{
plotVario
}
\details{
description, a paragraph
}
\examples{
# not run
}
