% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotZ}
\alias{plotZ}
\title{plotSp}
\usage{
plotZ(Z, map = NULL, id = FALSE)
}
\arguments{
\item{Z}{xxxx}

\item{map}{xxxx}

\item{id}{xxxx}
}
\value{
a plot
}
\description{
plotSp
}
\details{
description, a paragraph
}
\examples{
# not run
}
