% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotzf}
\alias{plotzf}
\title{plotzf}
\usage{
plotzf(qProb, map, pdf1 = NULL, FULL = F, i, vecj, data = NULL)
}
\arguments{
\item{qProb}{xxxx}

\item{map}{xxxx}

\item{pdf1}{xxxx}

\item{FULL}{xxxx}

\item{i}{xxxx}

\item{vecj}{xxxx}

\item{data}{xxxx}
}
\value{
a plot
}
\description{
plotzf
}
\details{
description, a paragraph Attention, voir la sortie de cette fonction en package!!!
}
\examples{
# not run
}
