% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calCost.R
\name{Points_Near_Boundary}
\alias{Points_Near_Boundary}
\title{Points_Near_Boundary}
\usage{
Points_Near_Boundary(map)
}
\arguments{
\item{map}{object returned by function genMap or genMapR}
}
\description{
Points_Near_Boundary
}
\details{
function that returns a list of points in a zone that are near boundary of the map
}
\examples{
map = mapTest
Points_Near_Boundary(map = map)
}
