% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{pointsSp}
\alias{pointsSp}
\title{pointsSp}
\usage{
pointsSp(sp, k = 1, col = "red")
}
\arguments{
\item{sp}{SpatialPolygons object}

\item{k}{polygon number}

\item{col}{color}
}
\description{
pointsSp
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
plotZ(Z)
pointsSp(Z[[1]])
# not run
}
