% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{polyToSp}
\alias{polyToSp}
\title{polyToSp}
\usage{
polyToSp(Z, iZ, k)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iZ}{zone number}

\item{k}{polygon number}
}
\value{
a SpatialPolygons
}
\description{
polyToSp
}
\details{
transforms kth polygon of zone into SpatialPolygons
}
\examples{
ZK=initialZoning(qProb=c(0.4,0.2,0.7),mapTest)
Z=ZK$resZ$zonePolygone
sph=geozoning:::polyToSp(Z,5,2)
plotZ(Z)
sp::plot(sph,type="l",col="blue",add=TRUE)
# not run
}
\keyword{internal}
