% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{polyToSp2}
\alias{polyToSp2}
\title{polyToSp2}
\usage{
polyToSp2(p)
}
\arguments{
\item{p}{polygon}
}
\value{
a SpatialPolygons
}
\description{
polyToSp2
}
\details{
transforms polygon into SpatialPolygons
}
\examples{
ZK=initialZoning(qProb=c(0.4,0.2,0.7),mapTest)
Z=ZK$resZ$zonePolygone
sp=Z[[5]]
P1=geozoning:::getPolySp(sp,1)
sph=polyToSp2(P1)
plotZ(Z)
sp::plot(sph,col="blue",lwd=2,add=TRUE)
}
