% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{printInterZ}
\alias{printInterZ}
\title{printInterZ}
\usage{
printInterZ(Z, sp)
}
\arguments{
\item{Z}{xxxx}

\item{sp}{xxxx}
}
\value{
a message?
}
\description{
printInterZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
