% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{printLabZ}
\alias{printLabZ}
\title{printLabZ}
\usage{
printLabZ(Klist)
}
\arguments{
\item{Klist}{list of zoning objects, typically result of a call to correctionTree}
}
\value{
a list of zoning objects
}
\description{
printLabZ
}
\details{
print zoning labels for a list of zoning objects
}
\examples{
data(mapTest)
qProb=c(0.1,0.2,0.4);criti=correctionTree(qProb,mapTest) # 2 zonings at last level
printLabZ(criti$zk[[2]])
# not run
}
