% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{printZid}
\alias{printZid}
\title{printZid}
\usage{
printZid(Z)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}
}
\description{
printZid
}
\details{
print zone identifiers in a zoning
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
printZid(Z)
}
