% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{ptInZone}
\alias{ptInZone}
\title{ptInZone}
\usage{
ptInZone(zone, pts, numpt)
}
\arguments{
\item{zone}{SpatialPolygons}

\item{pts}{data points}

\item{numpt}{data point indices}
}
\value{
1 if point is within zone, 0 if not
}
\description{
ptInZone
}
\details{
description, a paragraph
}
\examples{
data(mapTest)
data(resZTest)
Z=resZTest$zonePolygone
ptInZone(Z[[1]],mapTest$krigData,c(5,500))
# not run
}
