% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRandKmap.R
\name{ptNei}
\alias{ptNei}
\title{returns list of point neigbors for each point}
\usage{
ptNei(neighBool)
}
\arguments{
\item{neighBool}{numeric, boolean neighborhood matrix for pts}
}
\value{
a list of pt neigbors for each pt
}
\description{
returns list of point neigbors for each point
}
\examples{
# not run, take a while...
#data(mapTest) # simulated data
#grid=genEmptyGrid(calStep(2000,1,1),1,1)
#nbP= grid$nx*grid$ny
#neighBool=matrix(logical(nbP^2),nbP,nbP)
#resVoronoi=voronoiPolygons(mapTest$krigData,c(0,1,0,1),neighBool)
#neighBool=resVoronoi$neighBool
#listeNpt=ptNei(neighBool)
}
