% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ptsInSp}
\alias{ptsInSp}
\title{ptInZone}
\usage{
ptsInSp(sp, pts, hole = FALSE)
}
\arguments{
\item{sp}{xxxx}

\item{pts}{xxxx}

\item{hole}{xxxx}
}
\value{
a ?
}
\description{
ptInZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
