% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{r2}
\alias{r2}
\title{r2}
\usage{
r2(reslm)
}
\arguments{
\item{reslm}{result of a call to lm}
}
\value{
the adjusted r-square of the lm model
}
\description{
r2
}
\details{
adjusted R2
}
\examples{
# not run
}
