% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randKmap.R
\name{randKmap}
\alias{randKmap}
\title{randKmap:  Generate data for zoning or prepare real data}
\usage{
randKmap(DataObj, seed = NULL, nPoints = 450, nPointsK = 2000,
  nSimuCond = 0, typeMod = "Gau", Vpsill = 5, Vrange = 0.2, Vmean = 8,
  Vnugget = 0, boundary = list(x = c(0, 0, 1, 1, 0), y = c(0, 1, 1, 0, 0)),
  manualBoundary = FALSE, krig = 2, disp = 0, FULL = FALSE)
}
\arguments{
\item{DataObj}{=NULL: simulated data with seed or = a data frame with real data}

\item{seed}{numeric, seed}

\item{nPoints}{numeric, number of points, default 450}

\item{nPointsK}{numeric, default 2000}

\item{nSimuCond}{numeric}

\item{typeMod}{character, model type}

\item{Vpsill}{numeric, default 5}

\item{Vrange}{numeric, default 0.2}

\item{Vmean}{numeric, default 8}

\item{Vnugget}{numeric, default 0}

\item{boundary}{list contains x and y}

\item{manualBoundary}{logical, default FALSE}

\item{krig}{numeric}

\item{disp}{numeric}

\item{FULL}{logical, if TRUE the returned list is complete}
}
\value{
a list
\describe{
\item{rawData}{simulated or real raw data within the boundary}
\item{step}{grid step}
\item{krigData}{kriged data}
\item{krigGrid}{kriged data in form of grid}
\item{krigN}{kriged neighbours of each data point}
\item{krigSurfVoronoi}{areas of Voronoi polygons in the tesselation of kriged data}
\item{modelGen}{random fields model}
\item{VGMmodel}{vgm model}
\item{boundary}{(x,y) list of boundary points}
}
}
\description{
randKmap:  Generate data for zoning or prepare real data
}
\details{
generates a map object from simulated data or real data
}
\examples{
# not run
map = randKmap(NULL,nPointsK=500,Vmean=15,krig=2)
mean(map$krigGrid) # mean of generated kriged data
plotMap(map)

}
