% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readS.R
\name{readS}
\alias{readS}
\title{readS returns coords, ranges for x and y of a shapefile}
\usage{
readS(file, rep)
}
\arguments{
\item{file}{file name}

\item{rep}{directory}
}
\value{
a ?
}
\description{
readS returns coords, ranges for x and y of a shapefile
}
\details{
description, a paragraph
}
\examples{
# not run
}
