% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{remove1FromZ}
\alias{remove1FromZ}
\title{remove1FromZ}
\usage{
remove1FromZ(Z, iC, zoneN, simplitol, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{iC}{xxxx}

\item{zoneN}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
remove1FromZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
