% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resZTest}
\alias{resZTest}
\title{A list of initialZoning results}
\format{a list with all results from initialZoning function: criterion value, list with components matDistance, matDistanceCorr and cost, such as returned by a call to calDistance, a list with components zoneN, zoneNModif, listZonePoint, meanTot, meanZone,listSurf, critSurf, zonePolygone, such as the object returned by calNei.}
\usage{
resZTest
}
\description{
A list of initialZoning results
}
\keyword{datasets}
