% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{rowlings}
\alias{rowlings}
\title{rowlings}
\usage{
rowlings(pts, ply)
}
\arguments{
\item{pts}{xxxx}

\item{ply}{xxxx}
}
\value{
a ?
}
\description{
rowlings
}
\details{
description, a paragraph
}
\examples{
# not run
}
