% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{rowlings2}
\alias{rowlings2}
\title{rowlings2}
\usage{
rowlings2(pts, ply)
}
\arguments{
\item{pts}{xxxx}

\item{ply}{xxxx}
}
\value{
a ?
}
\description{
rowlings2
}
\details{
description, a paragraph
}
\examples{
# not run
}
