% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveZoningFromSimu.R
\name{saveZoningFromSimu}
\alias{saveZoningFromSimu}
\title{saveZoningFromSimu}
\usage{
saveZoningFromSimu(seed = 80, qProb = 0.4, disp = disp)
}
\arguments{
\item{seed}{xxxx}

\item{qProb}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
saveZoningFromSimu
}
\details{
description, a paragraph
}
\examples{
# not run
}
