% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{searchNODcrit}
\alias{searchNODcrit}
\title{searchNODcrit}
\usage{
searchNODcrit(qProb, le, zk, criterion, cost, costL, nz)
}
\arguments{
\item{qProb}{probability vector used to generate quantile values}

\item{le}{level index}

\item{zk}{list of zonings}

\item{criterion}{list of criteria}

\item{cost}{list of costs}

\item{costL}{list of per label costs}

\item{nz}{list of numbers of zones}
}
\value{
a list with components
\describe{
\item{ind}{index of last level zoning that has the higher criterion value}
\item{critList}{criterion value corresponding to best last level zoning}
\item{costlist}{cost value corresponding to best last level zoning}
\item{costLlist}{cost per label value corresponding to best last level zoning}
\item{nzList}{number of zones of best last level zoning}
\item{nq}{lenght of quantile vector}
}
}
\description{
searchNODcrit
}
\details{
description, a paragraph
}
\examples{
qProb=c(0.1,0.2);criti=correctionTree(qProb,mapTest)
res=geozoning:::searchNODcrit(qProb,2,criti$zk,criti$criterion,criti$cost,criti$costL,criti$nz)
# not run
}
\keyword{internal}
