% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{searchNODcrit1}
\alias{searchNODcrit1}
\title{searchNODcrit1}
\usage{
searchNODcrit1(qProb, crit)
}
\arguments{
\item{qProb}{probability vector used to generate quantile values}

\item{crit}{result of call to correctionTree (with SAVE=TRUE)}
}
\value{
a list with components
\describe{
\item{ind}{index of last level zoning that has the higher criterion value}
\item{critList}{criterion value corresponding to best last level zoning}
\item{costlist}{cost value corresponding to best last level zoning}
\item{costLlist}{cost per label value corresponding to best last level zoning}
\item{nzList}{number of zones of best last level zoning}
}
}
\description{
searchNODcrit1
}
\details{
description, a paragraph
}
\examples{
data(mapTest)
qProb=c(0.1,0.2,0.4);criti=correctionTree(qProb,mapTest) # 2 zonings at last level
res=searchNODcrit1(qProb,criti)# best one is frist element of last level
# not run
}
