% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selMaps.R
\name{selMaps}
\alias{selMaps}
\title{selMaps}
\usage{
selMaps(seed = 89, thr = 0.5, med = NULL, medC = NULL, m1 = NULL,
  m2 = NULL, m3 = NULL, m4 = NULL, m5 = NULL)
}
\arguments{
\item{seed}{xxxx}

\item{thr}{xxxx}

\item{med}{xxxx}

\item{medC}{xxxx}

\item{m1}{xxxx}

\item{m2}{xxxx}

\item{m3}{xxxx}

\item{m4}{xxxx}

\item{m5}{xxxx}
}
\value{
a ?
}
\description{
selMaps
}
\details{
description, a paragraph
}
\examples{
# not run
}
