% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{separationPoly}
\alias{separationPoly}
\title{separationPoly}
\usage{
separationPoly(polyTot)
}
\arguments{
\item{polyTot}{xxxx}
}
\value{
a ?
}
\description{
separationPoly
}
\details{
description, a paragraph
}
\examples{
# not run
}
