% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{setId}
\alias{setId}
\title{setId}
\usage{
setId(Z, iZ, id)
}
\arguments{
\item{Z}{xxxx}

\item{iZ}{xxxx}

\item{id}{xxxx}
}
\value{
a ?
}
\description{
setId
}
\details{
description, a paragraph
}
\examples{
# not run
}
