% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape1}
\alias{shape1}
\title{An external zoning read from a shape file}
\format{a SpatialPolygons object:}
\usage{
shape1
}
\description{
An external zoning read from a shape file
}
\keyword{datasets}
