% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{Sigmai2}
\alias{Sigmai2}
\title{Sigmai2}
\usage{
Sigmai2(index, listZonePoint, tabVal, surfaceVoronoi, meanZone2)
}
\arguments{
\item{index}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfaceVoronoi}{xxxx}

\item{meanZone2}{xxxx}
}
\value{
a ?
}
\description{
Sigmai2
}
\details{
description, a paragraph
}
\examples{
# not run
}
