% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{SigmaL2}
\alias{SigmaL2}
\title{SigmaL2}
\usage{
SigmaL2(zlab, listZonePoint, tabVal, surfVoronoi)
}
\arguments{
\item{zlab}{xxxx}

\item{listZonePoint}{xxxx}

\item{tabVal}{xxxx}

\item{surfVoronoi}{xxxx}
}
\value{
a ?
}
\description{
SigmaL2
}
\details{
description, a paragraph
}
\examples{
# not run
}
