% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuCond.R
\name{simulCond}
\alias{simulCond}
\title{simulCond}
\usage{
simulCond(step, tabAlea, dosstest, nbSimul, modelVar, bordure)
}
\arguments{
\item{step}{xxxx}

\item{tabAlea}{xxxx}

\item{dosstest}{xxxx}

\item{nbSimul}{xxxx}

\item{modelVar}{xxxx}

\item{bordure}{xxxx}
}
\value{
a ?
}
\description{
simulCond
}
\details{
description, a paragraph
}
\examples{
# not run
}
