% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothingMap.R
\name{smoothingMap}
\alias{smoothingMap}
\title{smoothingMap}
\usage{
smoothingMap(Z, width = 0.01, map, disp = FALSE)
}
\arguments{
\item{Z}{list of zone}

\item{width}{smoothing parameter}

\item{map}{object returned by function genMap}

\item{disp}{logical, if TRUE display the successful step of the program, otherwise do not display}
}
\value{
a new list of smoothed zones.
}
\description{
smoothingMap
}
\details{
function that smooths all zones of map
}
