% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortCrit.R
\name{sortCrit}
\alias{sortCrit}
\title{sortCrit}
\usage{
sortCrit(qProb, crit, cost, costL, nz, mdist, listOfZ, map, disp,
  SAVE = FALSE)
}
\arguments{
\item{qProb}{xxxx}

\item{crit}{xxxx}

\item{cost}{xxxx}

\item{costL}{xxxx}

\item{nz}{xxxx}

\item{mdist}{xxxx}

\item{listOfZ}{xxxx}

\item{map}{xxxx}

\item{disp}{xxxx}

\item{SAVE}{xxxx}
}
\value{
a ?
}
\description{
sortCrit
}
\details{
description, a paragraph
}
\examples{
# not run
}
