% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{spnorm}
\alias{spnorm}
\title{spnorm}
\usage{
spnorm(p, boundary)
}
\arguments{
\item{p}{xxxx}

\item{boundary}{xxxx}
}
\value{
a list
}
\description{
spnorm
}
\details{
description, a paragraph
}
\examples{
# not run
}
