% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{spToSL}
\alias{spToSL}
\title{spToSL}
\usage{
spToSL(sp)
}
\arguments{
\item{sp}{SpatialPolygons}
}
\value{
a SpatialLines
}
\description{
spToSL
}
\details{
tranform SpatialPolygons into SpatialLines
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
spToSL(Z[[5]])
}
