% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCriteria.R
\name{studyCriteria}
\alias{studyCriteria}
\title{studyCriteria}
\usage{
studyCriteria(vseed = 89)
}
\arguments{
\item{vseed}{xxxx}
}
\value{
a ?
}
\description{
studyCriteria
}
\details{
description, a paragraph
}
\examples{
# not run
}
