% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCriteria.R
\name{studyCriteria2}
\alias{studyCriteria2}
\title{studyCriteria2}
\usage{
studyCriteria2(vseed)
}
\arguments{
\item{vseed}{xxxx}
}
\value{
a ?
}
\description{
studyCriteria2
}
\details{
description, a paragraph
}
\examples{
# not run
}
