% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCriteria.R
\name{studyCriteria3}
\alias{studyCriteria3}
\title{studyCriteria3}
\usage{
studyCriteria3(vseed = NULL, thr = 0.75, med = NULL, medM = NULL)
}
\arguments{
\item{vseed}{xxxx}

\item{thr}{xxxx}

\item{med}{xxxx}

\item{medM}{xxxx}
}
\value{
a ?
}
\description{
studyCriteria3
}
\details{
description, a paragraph
}
\examples{
# not run
}
