% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCriteria.R
\name{studyCriteria4}
\alias{studyCriteria4}
\title{studyCriteria4}
\usage{
studyCriteria4(vseed = NULL, thr = 0.75, med = NULL, medC = NULL)
}
\arguments{
\item{vseed}{xxxx}

\item{thr}{xxxx}

\item{med}{xxxx}

\item{medC}{xxxx}
}
\value{
a ?
}
\description{
studyCriteria4
}
\details{
description, a paragraph
}
\examples{
# not run
}
