% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCriteria.R
\name{studyCriteria5}
\alias{studyCriteria5}
\title{studyCriteria5}
\usage{
studyCriteria5(map, m1, m2, m3, m4, m5, title = "")
}
\arguments{
\item{map}{xxxx}

\item{m1}{xxxx}

\item{m2}{xxxx}

\item{m3}{xxxx}

\item{m4}{xxxx}

\item{m5}{xxxx}

\item{title}{xxxx}
}
\value{
a ?
}
\description{
studyCriteria5
}
\details{
description, a paragraph
}
\examples{
# not run
}
