% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{superLines}
\alias{superLines}
\title{superLines}
\usage{
superLines(boundary)
}
\arguments{
\item{boundary}{list, contains x and y coordinates of map boundaries}
}
\value{
a SpatialLines object
}
\description{
superLines
}
\details{
converts boundary (list of x and y pts) into Spatial Lines
}
\examples{
data(mapTest)
superL=superLines(mapTest$boundary)
sp::plot(superL)
}
