% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{testInterSpe}
\alias{testInterSpe}
\title{testInterSpe}
\usage{
testInterSpe(Z, i1, i2)
}
\arguments{
\item{Z}{xxxx}

\item{i1}{xxxx}

\item{i2}{xxxx}
}
\value{
a ?
}
\description{
testInterSpe
}
\details{
description, a paragraph
}
\examples{
# not run
}
