% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{testInterSpeZ}
\alias{testInterSpeZ}
\title{testInterSpeZ}
\usage{
testInterSpeZ(Z)
}
\arguments{
\item{Z}{xxxx}
}
\value{
a ?
}
\description{
testInterSpeZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
