% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{testInterSpeZ1}
\alias{testInterSpeZ1}
\title{testInterSpeZ1}
\usage{
testInterSpeZ1(Z, i1)
}
\arguments{
\item{Z}{xxxx}

\item{i1}{xxxx}
}
\value{
a ?
}
\description{
testInterSpeZ1
}
\details{
description, a paragraph
}
\examples{
# not run
}
