% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{transfoSpPoly}
\alias{transfoSpPoly}
\title{transfoSpPoly}
\usage{
transfoSpPoly(polyL)
}
\arguments{
\item{polyL}{xxxx}
}
\value{
a ?
}
\description{
transfoSpPoly
}
\details{
description, a paragraph
}
\examples{
# not run
}
