% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{trLabZone}
\alias{trLabZone}
\title{trLabZone}
\usage{
trLabZone(K1, K2, Z1, Z2, map, qProb, disp = 0)
}
\arguments{
\item{K1}{xxxx}

\item{K2}{xxxx}

\item{Z1}{xxxx}

\item{Z2}{xxxx}

\item{map}{xxxx}

\item{qProb}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
trLabZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
