% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateZK.R
\name{updateZK}
\alias{updateZK}
\title{updateZK}
\usage{
updateZK(map, qProb, le, kk, listOfZ, crit, cost, costL, nz, mdist, K1, Z1, Z2,
  pErr, optiCrit, simplitol)
}
\arguments{
\item{map}{xxxx}

\item{qProb}{xxxx}

\item{le}{xxxx}

\item{kk}{xxxx}

\item{listOfZ}{xxxx}

\item{crit}{xxxx}

\item{cost}{xxxx}

\item{costL}{xxxx}

\item{nz}{xxxx}

\item{mdist}{xxxx}

\item{K1}{xxxx}

\item{Z1}{xxxx}

\item{Z2}{xxxx}

\item{pErr}{xxxx}

\item{optiCrit}{xxxx}

\item{simplitol}{xxxx}
}
\value{
a ?
}
\description{
updateZK
}
\details{
description, a paragraph
}
\examples{
# not run
}
