% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{valZ}
\alias{valZ}
\title{valZ}
\usage{
valZ(map, K)
}
\arguments{
\item{map}{map object returned by genMap function}

\item{K}{zoning object (such as returned by calNei function)}
}
\value{
a list with components
\describe{
\item{val}{list with vector of data values for each zone, zones are sorted by increasing mean values}
\item{ord}{order of zones sorted by increasing mean values}
}
}
\description{
valZ
}
\details{
sorts zones according to attribute mean value
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.7),mapTest,SAVE=TRUE) 
K=criti$zk[[2]][[1]]
valZ(mapTest,K)
# not run
}
