% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visudispZ.R
\name{visuZ}
\alias{visuZ}
\title{visuZ}
\usage{
visuZ(qProb, map, nq = 3, zk, disp = TRUE)
}
\arguments{
\item{qProb}{xxxx}

\item{map}{xxxx}

\item{nq}{xxxx}

\item{zk}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
visuZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
