% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalNei.R
\name{voronoiPolygons}
\alias{voronoiPolygons}
\title{voronoiPolygons}
\usage{
voronoiPolygons(spdata, neighBool, PTJUNCTION = FALSE, FULL = FALSE)
}
\arguments{
\item{spdata}{xxxx}

\item{neighBool}{xxxx}

\item{PTJUNCTION}{xxxx}

\item{FULL}{xxxx}
}
\value{
a ?
}
\description{
voronoiPolygons
}
\details{
description, a paragraph
}
\examples{
# not run
}
\seealso{
http://www.carsonfarmer.com/2009/09/voronoi-polygons-with-r/
}
