% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yield}
\alias{yield}
\title{A data frame with real data used for zoning}
\format{a data frame with 6415 rows and 3 variables:
\describe{
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{Yield}{numeric variable - phenotype}
}}
\usage{
yield
}
\description{
A data frame with real data used for zoning
}
\keyword{datasets}
