% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ZnormXY}
\alias{ZnormXY}
\title{normalize polygons in zoning}
\usage{
ZnormXY(Z, xmin, xmax, ymin, ymax)
}
\arguments{
\item{Z}{xxxx}

\item{xmin}{xxxx}

\item{xmax}{xxxx}

\item{ymin}{xxxx}

\item{ymax}{xxxx}
}
\value{
a dataframe?
}
\description{
normalize polygons in zoning
}
\details{
description, a paragraph
}
\examples{
# not run
}
