% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{zoneAssign}
\alias{zoneAssign}
\title{zoneAssign}
\usage{
zoneAssign(tab, Z)
}
\arguments{
\item{tab}{xxxx}

\item{Z}{xxxx}
}
\value{
a ?
}
\description{
zoneAssign
}
\details{
description, a paragraph
}
\examples{
# not run
}
