% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{zoneFusion3}
\alias{zoneFusion3}
\title{zoneFusion3}
\usage{
zoneFusion3(Z, K, iC, Ns, map, minSize, simplitol, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{iC}{xxxx}

\item{Ns}{xxxx}

\item{map}{xxxx}

\item{minSize}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
zoneFusion3
}
\details{
description, a paragraph
}
\examples{
# not run
}
