% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{zoneFusion4}
\alias{zoneFusion4}
\title{zoneFusion4}
\usage{
zoneFusion4(Z, iSmall, iBig, simplitol = 0.001, disp = 0)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iSmall}{index of zone to remove by merging it into other zone}

\item{iBig}{index of zone to merge into}

\item{simplitol}{tolerance for spatial polygons geometry simplification}

\item{disp}{0: no info, 1: some info}
}
\value{
a new zoning geometry
}
\description{
zoneFusion4
}
\details{
merge 2 zones from given zoning
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
zoneFusion4(Z,5,4,disp=2)
}
