% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{zoneGeneration}
\alias{zoneGeneration}
\title{zoneGeneration}
\usage{
zoneGeneration(map, qProb = c(0.25, 0.75), GridData = FALSE)
}
\arguments{
\item{map}{xxxx}

\item{qProb}{xxxx}

\item{GridData}{xxxx}
}
\value{
a ?
}
\description{
zoneGeneration
}
\details{
description, a paragraph
}
\examples{
# not run
}
