% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{zoneGrow}
\alias{zoneGrow}
\title{zoneGrow}
\usage{
zoneGrow(Z, K, iC, Ns, map, optiCrit, valRef, qProb, minSizeNG, distIsoZ, LEQ,
  MAXP, simplitol, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{iC}{xxxx}

\item{Ns}{xxxx}

\item{map}{xxxx}

\item{optiCrit}{xxxx}

\item{valRef}{xxxx}

\item{qProb}{xxxx}

\item{minSizeNG}{xxxx}

\item{distIsoZ}{xxxx}

\item{LEQ}{xxxx}

\item{MAXP}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
zoneGrow
}
\details{
description, a paragraph
}
\examples{
# not run
}
