% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modNIso.R
\name{zoneModifnonIso}
\alias{zoneModifnonIso}
\title{zoneModifnonIso}
\usage{
zoneModifnonIso(Z, K, qProb, map, zoneClose, iC, simplitol, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{qProb}{xxxx}

\item{map}{xxxx}

\item{zoneClose}{xxxx}

\item{iC}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
zoneModifnonIso
}
\details{
description, a paragraph
}
\examples{
# not run
}
