% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoneQ.R
\name{zoneQ}
\alias{zoneQ}
\title{zoneQ}
\usage{
zoneQ(contourSp, iC, iE, Z, K, map, simplitol)
}
\arguments{
\item{contourSp}{xxxx}

\item{iC}{xxxx}

\item{iE}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}

\item{map}{xxxx}

\item{simplitol}{xxxx}
}
\value{
a ?
}
\description{
zoneQ
}
\details{
description, a paragraph
}
\examples{
# not run
}
