% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{addContour}
\alias{addContour}
\title{addContour}
\usage{
addContour(map, val, col = "blue", super = T)
}
\arguments{
\item{map}{xxxx}

\item{val}{xxxx}

\item{col}{xxxx}

\item{super}{xxxx}
}
\value{
a ?
}
\description{
addContour
}
\details{
description, a paragraph
}
\examples{
# not run
}
