% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalNei.R
\name{calcCritNarrow}
\alias{calcCritNarrow}
\title{detection of narrow zones (ratio area/perimeter^2)}
\usage{
calcCritNarrow(zonePolygone)
}
\arguments{
\item{zonePolygone}{zoning}
}
\value{
a numerical value
}
\description{
detection of narrow zones (ratio area/perimeter^2)
}
\details{
computes for each zone of a zoning the ratio area/squared perimeter
}
\examples{
data(resZTest)
calcCritNarrow(resZTest$zonePolygone)
}
