% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit1}
\alias{calCrit1}
\title{calCrit1}
\usage{
calCrit1(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{zone distance matrix resulting from a call to calDistance}

\item{zoneNModif}{matrix of zone neigbors with FALSE on the diagonal}
}
\value{
a numerical value
}
\description{
calCrit1
}
\details{
computes a quality criterion equal to min(mean(dij^2/(dii^2+dij^2)))
see also  \code{\link{calCrit2, calCrit3, calCrit4, calCrit5, calCritMinMean}} for other criteria
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
resD = calDistance(typedist=1,mapTest$krigData,K$listZonePoint,K$zoneN,
       mapTest$krigSurfVoronoi,K$meanZone,pErr=0.9)
calCrit1(resD$matDistanceCorr,K$zoneNModif)
# not run
}
