% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit1}
\alias{calCrit1}
\title{calCrit1}
\usage{
calCrit1(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCrit1
}
\details{
description, a paragraph
}
\examples{
# not run
}
