% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit3}
\alias{calCrit3}
\title{calCrit3}
\usage{
calCrit3(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCrit3
}
\details{
description, a paragraph
}
\examples{
# not run
}
