% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalCrit.R
\name{calCrit7}
\alias{calCrit7}
\title{calCrit7}
\usage{
calCrit7(matDistance, zoneNModif)
}
\arguments{
\item{matDistance}{xxxx}

\item{zoneNModif}{xxxx}
}
\value{
a ?
}
\description{
calCrit7
}
\details{
description, a paragraph
}
\examples{
# not run
}
