% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDistance.R
\name{calDistance}
\alias{calDistance}
\title{calDistance}
\usage{
calDistance(typedist, tabVal, listZonePoint, zoneN, surfVoronoi, meanZone, pErr)
}
\arguments{
\item{typedist}{xxxx}

\item{tabVal}{xxxx}

\item{listZonePoint}{xxxx}

\item{zoneN}{xxxx}

\item{surfVoronoi}{xxxx}

\item{meanZone}{xxxx}

\item{pErr}{xxxx}
}
\value{
a ?
}
\description{
calDistance
}
\details{
description, a paragraph
}
\examples{
# not run
}
