% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFrame.R
\name{calFrame}
\alias{calFrame}
\title{calFrame}
\usage{
calFrame(iC, Z, K, distIsoZ = 0.075)
}
\arguments{
\item{iC}{xxxx}

\item{Z}{xxxx}

\item{K}{xxxx}

\item{distIsoZ}{xxxx}
}
\value{
a ?
}
\description{
calFrame
}
\details{
description, a paragraph
}
\examples{
# not run
}
