% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calGearyGlo}
\alias{calGearyGlo}
\title{calGearyGlo}
\usage{
calGearyGlo(matVoisin, vectMoy, moyTot, vectSurface)
}
\arguments{
\item{matVoisin}{xxxx}

\item{vectMoy}{xxxx}

\item{moyTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
calGearyGlo
}
\details{
description, a paragraph
}
\examples{
# not run
}
