% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calGearyLoc}
\alias{calGearyLoc}
\title{local Geary criteria}
\usage{
calGearyLoc(matN, vectMean, meanTot, vectSurface)
}
\arguments{
\item{matN}{neighborhood (zone or point) matrix}

\item{vectMean}{vector of mean zone values}

\item{meanTot}{global mean}

\item{vectSurface}{vector of zone areas}
}
\value{
a vector of local Geary criteria
}
\description{
local Geary criteria
}
\details{
computes local Geary indices
}
\examples{
K=resZTest
zoneA=sapply(K$zonePolygone,rgeos::gArea)
calGearyLoc(K$zoneNModif,K$meanZone,K$meanTot,zoneA)
# not run
}
