% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranBLocal}
\alias{calMoranBLocal}
\title{compute local Moran indices (per zone)}
\usage{
calMoranBLocal(NZone, matDistanceMoranB, vectSurface)
}
\arguments{
\item{NZone}{xxxx}

\item{matDistanceMoranB}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
compute local Moran indices (per zone)
}
\details{
description, a paragraph
}
\examples{
# not run
}
