% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranBTot}
\alias{calMoranBTot}
\title{fonction qui compute un critère de moran sur toute la carte}
\usage{
calMoranBTot(voisinZone, matDistanceMoranB, vectSurface)
}
\arguments{
\item{voisinZone}{xxxx}

\item{matDistanceMoranB}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
fonction qui compute un critère de moran sur toute la carte
}
\details{
description, a paragraph
}
\examples{
# not run
}
