% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranGlo}
\alias{calMoranGlo}
\title{fonction calant un indice de Moran(adapté) entre les différentes zone,ou les différents points}
\usage{
calMoranGlo(matVoisinZone, vectMoy, moyTot, vectSurface)
}
\arguments{
\item{matVoisinZone}{xxxx}

\item{vectMoy}{xxxx}

\item{moyTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
fonction calant un indice de Moran(adapté) entre les différentes zone,ou les différents points
}
\details{
description, a paragraph
}
\examples{
# not run
}
