% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalIndices.R
\name{calMoranLoc}
\alias{calMoranLoc}
\title{calMoranLoc}
\usage{
calMoranLoc(matN, vectMean, meanTot, vectSurface)
}
\arguments{
\item{matN}{xxxx}

\item{vectMean}{xxxx}

\item{meanTot}{xxxx}

\item{vectSurface}{xxxx}
}
\value{
a ?
}
\description{
calMoranLoc
}
\details{
description, a paragraph
}
\examples{
# not run
}
