% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRMmodel.R
\name{calRMmodel}
\alias{calRMmodel}
\title{transform VGM model into model usable by RandomFields}
\usage{
calRMmodel(vgmodel)
}
\arguments{
\item{vgmodel}{à remplir}
}
\value{
anisotropy matrix
}
\description{
transform VGM model into model usable by RandomFields
}
\examples{
# not run
}
\seealso{
http://www.techmat.vgtu.lt/~art/proc/file/BudrLi.pdf
}
