% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRandKmap.R
\name{calStep}
\alias{calStep}
\title{compute step for square grid}
\usage{
calStep(nPointsK)
}
\arguments{
\item{nPointsK}{numeric}
}
\value{
a numeric
}
\description{
compute step for square grid
}
\examples{
# not run
}
