% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalNei.R
\name{calZoneN}
\alias{calZoneN}
\title{calZoneN}
\usage{
calZoneN(ptN, zoneN, listZonePoint)
}
\arguments{
\item{ptN}{xxxx}

\item{zoneN}{xxxx}

\item{listZonePoint}{xxxx}
}
\value{
a ?
}
\description{
calZoneN
}
\details{
description, a paragraph
}
\examples{
# not run
}
