% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contourArea}
\alias{contourArea}
\title{contourArea}
\usage{
contourArea(contour1)
}
\arguments{
\item{contour1}{xxxx}
}
\value{
a numeric?
}
\description{
contourArea
}
\details{
description, a paragraph
}
\examples{
# not run
}
