% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{contourAuto}
\alias{contourAuto}
\title{contourAuto}
\usage{
contourAuto(cL, step, matVal, vRef, boundary, GridData = FALSE)
}
\arguments{
\item{cL}{xxxx}

\item{step}{xxxx}

\item{matVal}{xxxx}

\item{vRef}{xxxx}

\item{boundary}{xxxx}

\item{GridData}{xxxx}
}
\value{
a ?
}
\description{
contourAuto
}
\details{
description, a paragraph
}
\examples{
# not run
}
