% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contourToSpp}
\alias{contourToSpp}
\title{contourToSpp}
\usage{
contourToSpp(contour1, step)
}
\arguments{
\item{contour1}{xxxx}

\item{step}{xxxx}
}
\value{
a numeric?
}
\description{
contourToSpp
}
\details{
description, a paragraph
}
\examples{
# not run
}
