% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNei.R
\name{correctN}
\alias{correctN}
\title{correctN}
\usage{
correctN(Z, zoneN)
}
\arguments{
\item{Z}{xxxx}

\item{zoneN}{xxxx}
}
\value{
a ?
}
\description{
correctN
}
\details{
description, a paragraph
}
\examples{
# not run
}
