% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calCost.R
\name{Cost_By_Laplace}
\alias{Cost_By_Laplace}
\title{Cost_By_Laplace}
\usage{
Cost_By_Laplace(map, Z, numZ, Estimation)
}
\arguments{
\item{map}{object returned by function genMap}

\item{Z}{: an example of zoning (a list of zones)}

\item{numZ}{: number of the zone in which the cost will be computed}

\item{Estimation}{: value of linear interpolation by solving Laplace's equation}
}
\value{
cost computed by replacing values in zone by linear interpolation
}
\description{
Cost_By_Laplace
}
\details{
: function that returns the criterion COST by approximating the valeur in a point of the grid by
           the linear interpolation (approximate solution of Laplace's equation. For more details see help of function
           Transition_Zone_Near_Boundary, Transition_Zone_Far_Boundary or Extreme_Zone)
}
\examples{
seed=2
map=genMap(DataObj=NULL,seed=seed,krig=2,typeMod="Gau")
ZK=initialZoning(qProb=c(0.55,0.85),map)
Z=ZK$resZ$zonePolygone # list of zones
lab = ZK$resZ$lab # label of zones
plotM(map = map,Z = Z,lab = lab, byLab = FALSE)
# zone 6 is a transition zone that has commun boundary with the map
numZ = 6
Estimation = Transition_Zone_Near_Boundary(map = map, Z = Z, numZ = numZ)
# compute the cost
cL = Cost_By_Laplace(map = map, Z = Z, numZ = numZ, Estimation = Estimation)
cM = Cost_By_Mean(map = map, Z = Z, numZ = numZ)
print(cL$cost_Laplace)
print(cM$cost_Mean)
# zone 6 is a zone with gradient
}
