% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calCost.R
\name{Cost_By_Mean}
\alias{Cost_By_Mean}
\title{Cost_By_Mean}
\usage{
Cost_By_Mean(map, Z, numZ)
}
\arguments{
\item{map}{object returned by function genMap}

\item{Z}{: an example of zoning (a list of zones)}

\item{numZ}{: number of the zone in which the cost will be computed}
}
\value{
the cost value as described
}
\description{
Cost_By_Mean
}
\details{
: function that returns the criterion COST by approximating the value at a point of the grid by
           the mean value of the zone.
}
\examples{
seed=2
map=genMap(DataObj=NULL,seed=seed,krig=2,typeMod="Gau")
ZK=initialZoning(qProb=c(0.55,0.85),map)
Z=ZK$resZ$zonePolygone # list of zones
lab = ZK$resZ$lab # label of zones
plotM(map = map,Z = Z,lab = lab, byLab = FALSE)
# zone 6 is a transition zone that has commun boundary with the map
numZ = 6
Estimation = Transition_Zone_Near_Boundary(map = map, Z = Z, numZ = numZ)
# compute the cost
cL = Cost_By_Laplace(map = map, Z = Z, numZ = numZ, Estimation = Estimation)
cM = Cost_By_Mean(map = map, Z = Z, numZ = numZ)
print(cL$cost_Laplace)
print(cM$cost_Mean)
# zone 6 is a zone with gradient
}
