% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{costLab}
\alias{costLab}
\title{costLab}
\usage{
costLab(K, map)
}
\arguments{
\item{K}{xxxx}

\item{map}{xxxx}
}
\value{
a ?
}
\description{
costLab
}
\details{
description, a paragraph
}
\examples{
# not run
}
