% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{createHoles}
\alias{createHoles}
\title{createHoles}
\usage{
createHoles(Z)
}
\arguments{
\item{Z}{list of zones, each zone is a SpatialPolygons}
}
\value{
a list of zones where holes are distinct SpatialPolygons
}
\description{
createHoles
}
\details{
description, a paragraph
}
\examples{
# not run
}
