% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{datanormXY}
\alias{datanormXY}
\title{normalize data coords}
\usage{
datanormXY(data)
}
\arguments{
\item{data}{frame with x and y components}
}
\value{
a normalized data frame
}
\description{
normalize data coords
}
\details{
normalize data coordinates between 0 and 1 with different ratios for x and y
}
\examples{
nPoints=500
x=runif(nPoints, min=0, max=1)
y=runif(nPoints, min=0, max=1)
range(x) # not [0,1]
tabData=data.frame(x=x,y=y)
tabData=geozoning:::datanormXY(tabData) # x,y ranges are now [0,1]
# not run
}
\keyword{internal}
