% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{detectSmallZones}
\alias{detectSmallZones}
\title{detectSmallZones}
\usage{
detectSmallZones(zonePolygone, minSize)
}
\arguments{
\item{zonePolygone}{list of zones, each zone is a SpatialPolygons}

\item{minSize}{zone area threshold under which a zone is too small to be manageable}
}
\value{
a vector of small zones indices
}
\description{
detectSmallZones
}
\details{
detect zones with area < minSize
}
\examples{
data(mapTest)
ZK=initialZoning(qProb=c(0.4,0.7),mapTest)
Z=ZK$resZ$zonePolygone
minSize=0.012
iSmall=detectSmallZones(Z,minSize) # 2 small zones 
# not run
}
