% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detZoneCE.R
\name{detZoneEng}
\alias{detZoneEng}
\title{detZoneEng}
\usage{
detZoneEng(iZ, Z, zoneN)
}
\arguments{
\item{iZ}{index of zone for which englobing zone is searched}

\item{Z}{zoning}

\item{zoneN}{modified zone neighborhood matrix (FALSE values on diagonal)}
}
\value{
an integer value (0 if no englobing zone was found, englobing zone index otherwise)
}
\description{
detZoneEng
}
\details{
description, a paragraph
}
\examples{
# load zoning results from test file
data(resZTest)
Z=resZTest$zonePolygone
zoneN=resZTest$zoneNModif
detZoneEng(3,Z,zoneN) # zone 2 englobes zone 3
detZoneEng(2,Z,zoneN) # no englobing zone for zone 2 
}
