% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{DIJ}
\alias{DIJ}
\title{DIJ}
\usage{
DIJ(i, j, sigmai2, meanZone2, pErr)
}
\arguments{
\item{i}{xxxx}

\item{j}{xxxx}

\item{sigmai2}{xxxx}

\item{meanZone2}{xxxx}

\item{pErr}{xxxx}
}
\value{
a ?
}
\description{
DIJ
}
\details{
description, a paragraph
}
\examples{
# not run
}
