% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDistance.R
\name{DIJ}
\alias{DIJ}
\title{DIJ}
\usage{
DIJ(i, j, sigmai2, meanZone, pErr)
}
\arguments{
\item{i}{zone index}

\item{j}{neighbor zone index}

\item{sigmai2}{vector of zone variances}

\item{meanZone}{list of zone mean values}

\item{pErr}{tolerance for distance correction}
}
\value{
a list with components d and dCorr
}
\description{
DIJ
}
\details{
description, a paragraph
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
nz=length(K$zonePolygone)
si2=rep(NA,nz)
for (kk in 1:nz){
si2[kk]=Sigmai2(kk,K$listZonePoint,mapTest$krigData,
        mapTest$krigSurfVoronoi,K$meanZone)$sigmai2
}
d12=DIJ(1,2,si2,K$meanzone,0.9)
# not run
}
