% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{distanceNormalisationSqrt}
\alias{distanceNormalisationSqrt}
\title{distanceNormalisationSqrt}
\usage{
distanceNormalisationSqrt(matDistance)
}
\arguments{
\item{matDistance}{xxxx}
}
\value{
a ?
}
\description{
distanceNormalisationSqrt
}
\details{
description, a paragraph
}
\examples{
# not run
}
