% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalDistance.R
\name{distanceNormalisationSum}
\alias{distanceNormalisationSum}
\title{distanceNormalisationSum}
\usage{
distanceNormalisationSum(matDistance)
}
\arguments{
\item{matDistance}{xxxx}
}
\value{
a ?
}
\description{
distanceNormalisationSum
}
\details{
description, a paragraph
}
\examples{
# not run
}
