% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{extensionLine}
\alias{extensionLine}
\title{extensionLine}
\usage{
extensionLine(contourL = NULL, step = NULL, bdSP, superLines)
}
\arguments{
\item{contourL}{contour line}

\item{step}{grid step as returned by calStep}

\item{bdSP}{list, contains x and y coordinates of map boundaries}

\item{superLines}{object returned by superLines(bdSP)}
}
\value{
a list
}
\description{
extensionLine
}
\details{
closes contour lines by extending them to their interesection with the map border
}
\examples{
data(mapTest)
step=mapTest$step
xsize=mapTest$xsize
ysize=mapTest$ysize
cL=contourLines(seq(step, xsize-step, by=step),seq(step, ysize-step, by=step),
               mapTest$krigGrid, levels = c(5,7))
plot(mapTest$boundary,type="l",col="red")
lines(cL[[1]])#contour line is not closed
lines(geozoning:::extensionLine(cL[[1]],step,sp::SpatialPoints(mapTest$boundary),
     geozoning:::superLines(mapTest$boundary)),col="red") #contour line is closed
}
\keyword{internal}
