% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractionPoly}
\alias{extractionPoly}
\title{extractionPoly}
\usage{
extractionPoly(polyTot)
}
\arguments{
\item{polyTot}{SpatialPolygons}
}
\value{
a list of SpatialPolygons
}
\description{
extractionPoly
}
\details{
extract all elements from SpatialPolygons, holes and full polygons are handled equally
}
\examples{
data(mapTest)
ZK=initialZoning(qProb=c(0.2,0.4,0.7),mapTest)
Z=ZK$resZ$zonePolygone
extractionPoly(Z[[5]]) # returns 2 SpatialPolygons
# not run
}
\keyword{internal}
