% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findN.R
\name{findN}
\alias{findN}
\title{findN}
\usage{
findN(Z, K, listN, iC, minSize = 0.012)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{listN}{xxxx}

\item{iC}{xxxx}

\item{minSize}{xxxx}
}
\value{
a ?
}
\description{
findN
}
\details{
description, a paragraph
}
\examples{
# not run
}
