% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{findNptInZone}
\alias{findNptInZone}
\title{findNptInZone}
\usage{
findNptInZone(K, i1, i2, map)
}
\arguments{
\item{K}{zoning object, as returned by the calNei function}

\item{i1}{first zone}

\item{i2}{second zone, where to search for neighbors of points in first zone}

\item{map}{object returned by function genMap}
}
\value{
a two-column matrix, the first column contains indices of pts in first zone which have at least one neighbor in second zone, the second column contains the neighbor indices.
}
\description{
findNptInZone
}
\details{
find, in a given zone, neighbor points of points belonging to another zone
}
\examples{
data(resZTest)
K=resZTest
# not run
}
