% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRandKmap.R
\name{genData}
\alias{genData}
\title{generate data}
\usage{
genData(DataObj = NULL, seed = 0, nPoints = 450, typeMod = "Gau",
  Vpsill = 5, Vrange = 0.2, Vmean = 8, Vnugget = 0, boundary = list(x
  = c(0, 0, 1, 1, 0), y = c(0, 1, 1, 0, 0)), manualBoundary = FALSE)
}
\arguments{
\item{DataObj}{=NULL: simulated data with seed or = a data frame with real data}

\item{seed}{numeric,}

\item{nPoints}{numeric,}

\item{typeMod}{character}

\item{Vpsill}{numeric,}

\item{Vrange}{numeric,}

\item{Vmean}{numeric,}

\item{Vnugget}{numeric,}

\item{boundary}{list, contains x and y}

\item{manualBoundary}{logical,}
}
\value{
a list
\describe{
\item{tabData}{tabData}
\item{boundary}{boundary}
\item{modelGen}{modelGen}
\item{modelVGM}{modelVGM}
}
}
\description{
generate data
}
\details{
description, a paragraph
}
\examples{
# not run
# resGene=genData(DataObj,0,450,"Gau",5,0.2,8,0,list(x=c(0,0,1,1,0),y=c(0,1,1,0,0)),FALSE)
}
