% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRandKmap.R
\name{genEmptyGrid}
\alias{genEmptyGrid}
\title{generate grid from raw data}
\usage{
genEmptyGrid(step, nKrigE)
}
\arguments{
\item{step}{numeric}

\item{nKrigE}{numeric}
}
\value{
a dataframe that contains kriged positions based on original ones
}
\description{
generate grid from raw data
}
\examples{
# not run
}
