% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMap.R
\name{genMap}
\alias{genMap}
\title{wrapper for randKmap, generate 2D map}
\usage{
genMap(DataObj = NULL, seed = 80, krig = 2, Vpsill = 5, Vrange = 0.2,
  Vnugget = 0.2, Vmean = 8, nPointsK = 1000, boundary = list(x = c(0, 0,
  1, 1, 0), y = c(0, 1, 1, 0, 0)), disp = 0, FULL = FALSE)
}
\arguments{
\item{DataObj}{=NULL: simulated data with seed or = a data frame with real data}

\item{seed}{numeric,}

\item{krig}{numeric, 1: kriging with vgm model, 2: inverse distance kriging}

\item{Vpsill}{numeric parameter of the variogram model,}

\item{Vrange}{numeric parameter of the variogram model,}

\item{Vnugget}{numeric parameter of the variogram model,}

\item{Vmean}{numeric parameter of the variogram model,}

\item{nPointsK}{number of generated points after kriging}

\item{boundary}{list, contains x and y coordinates of map boundaries}

\item{disp}{numeric,}

\item{FULL}{logical, if TRUE the returned list is complete}
}
\value{
a map object as a list with components
\describe{
\item{tabAlea}{raw data, SpatialPointsDataFrame}
\item{surfaceVoronoi}{Voronoi polygon surfaces}
\item{krigTabAlea}{kriged data, SpatialPointsDataFrame}
\item{fitVarioAlea}{variogram}
\item{DataObj}{DataObj}
}
}
\description{
wrapper for randKmap, generate 2D map
}
\details{
wrapper for randKmap, generate 2D map with 1000 kriged data points, Gaussian field
}
\examples{
m=genMap(seed=1,krig=2,disp=1) #generates a map and plots data
mean(m$krigGrid) # mean of generated kriged data
# not run
}
