% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{genQseq}
\alias{genQseq}
\title{genQseq}
\usage{
genQseq(qProb, K, map, i1, i2, LEQ = 5, MAXP = 0.1, disp = 0)
}
\arguments{
\item{qProb}{xxxx}

\item{K}{xxxx}

\item{map}{xxxx}

\item{i1}{current zone index}

\item{i2}{englobing zone index}

\item{LEQ}{xxxx}

\item{MAXP}{xxxx}

\item{disp}{xxxx}
}
\value{
a plot
}
\description{
genQseq
}
\details{
description, a paragraph
}
\examples{
# not run
}
