% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getClosePt}
\alias{getClosePt}
\title{getClosePt}
\usage{
getClosePt(Z, indiceCourant, indiceZP, disp = FALSE)
}
\arguments{
\item{Z}{xxxx}

\item{indiceCourant}{xxxx}

\item{indiceZP}{xxxx}

\item{disp}{xxxx}
}
\value{
a numeric?
}
\description{
getClosePt
}
\details{
description, a paragraph
}
\examples{
# not run
}
