% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getCoords}
\alias{getCoords}
\title{getCoords}
\usage{
getCoords(sp, k = 1)
}
\arguments{
\item{sp}{xxxx}

\item{k}{xxxxxx}
}
\value{
some coordinates
}
\description{
getCoords
}
\details{
description, a paragraph
}
\examples{
# not run
}
