% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getIds}
\alias{getIds}
\title{getIds}
\usage{
getIds(Z, nums = NULL)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{nums}{zone numbers}
}
\value{
a character vector giving the zone identifiers
}
\description{
getIds
}
\details{
get zone identifiers in a zoning
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.5),mapTest,SAVE=TRUE)
Z=criti$zk[[2]][[1]]$zonePolygon
geozoning:::getIds(Z)
# not run
}
\keyword{internal}
