% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getNs}
\alias{getNs}
\title{getNs}
\usage{
getNs(zoneN, iZ)
}
\arguments{
\item{zoneN}{zone neighborhood Logical matrix}

\item{iZ}{index of current zone in zoning}
}
\value{
a Logical vector of current zone neighbors
}
\description{
getNs
}
\details{
get zone numbers of neighbors of a given zone
}
\examples{
data(mapTest)
K=resZTest
Ns=getNs(K$zoneNModif,5) # find neighbors of zone 5
}
\keyword{internal}
