% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getNumZone}
\alias{getNumZone}
\title{getNumZone}
\usage{
getNumZone(ptsp, Z)
}
\arguments{
\item{ptsp}{SpatialPointsDataFrame}

\item{Z}{zoning geometry (list of SpatialPolygons)}
}
\value{
the zone number
}
\description{
getNumZone
}
\details{
get zone numbers to which each point in a SpatialPointsDataFrame belongs
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
Z=K$zonePolygone
getNumZone(mapTest$krigData,Z)
}
