% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getPoly}
\alias{getPoly}
\title{getPoly}
\usage{
getPoly(Z, iZ, k)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iZ}{current zone index}

\item{k}{polygon number within current zone}
}
\value{
a polygon (object of class Polygon)
}
\description{
getPoly
}
\details{
get the kth polygon of the current zone in zoning Z
}
\examples{
ZK=initialZoning(qProb=c(0.4,0.2,0.7),mapTest)
Z=ZK$resZ$zonePolygone
P1=getPoly(Z,5,1)
P2=getPoly(Z,5,2) # second polygon is a hole
plot(P1@coords,type="l")
lines(P2@coords,type="l",col="blue")
# not run
}
