% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{getSurf}
\alias{getSurf}
\title{getSurf}
\usage{
getSurf(Z, iZ)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iZ}{zone number}
}
\value{
zone area
}
\description{
getSurf
}
\details{
get surface area of a given zone within a zoning
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
getSurf(Z,1)
# not run
}
\keyword{internal}
