% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridXY.R
\name{gridXY}
\alias{gridXY}
\title{generate empty grid}
\usage{
gridXY(mat)
}
\arguments{
\item{mat}{xxxx}
}
\value{
a ?
}
\description{
generate empty grid
}
\details{
description, a paragraph
}
\examples{
# not run
}
