% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{Identify}
\alias{Identify}
\title{Identify}
\usage{
Identify(id, Z)
}
\arguments{
\item{id}{zone identifier (character vector)}

\item{Z}{zoning geometry (list of SpatialPolygons)}
}
\value{
the zone number
}
\description{
Identify
}
\details{
get the number of a zone with a given identifier in a zoning
this is necessary because correction procedures may remove zones from initial#' zoning. Therefore zone numbers change, but identifiers are conserved.
}
\examples{
\donttest{
data(mapTest)
criti=correctionTree(c(0.4,0.5),mapTest,SAVE=TRUE)
Z=criti$zk[[2]][[1]]$zonePolygon
Identify(6,Z)
}
}
