% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interCB}
\alias{interCB}
\title{interCB}
\usage{
interCB(contour1, step, bd = boundary, envel, disp = 0)
}
\arguments{
\item{contour1}{xxxx}

\item{step}{xxxx}

\item{bd}{xxxx}

\item{envel}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
interCB
}
\details{
description, a paragraph
}
\examples{
# not run
}
