% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{interCB}
\alias{interCB}
\title{interCB}
\usage{
interCB(co, step, bd = list(x = c(0, 0, 1, 1, 0), y = c(0, 1, 1, 0, 0)),
  envel, disp = 0)
}
\arguments{
\item{co}{contour line}

\item{step}{map grid resolution}

\item{bd}{map boundary}

\item{envel}{envelope}

\item{disp}{info level (0-no info, 1- add lines to plot)}
}
\value{
a SpatialPolygons
}
\description{
interCB
}
\details{
generates SpatialPolygons object corresponding to intersection of contour with boundary, must be within SpatialPolygons given in envel argument
}
\examples{
data(mapTest)
pG=polyToSp2(sp::Polygon(mapTest$boundary)) #SpatialPolygons corresponding to map boundary
cL=contourAuto(list(),mapTest$step,mapTest$xsize,mapTest$ysize,
   mapTest$krigGrid,c(5,7),mapTest$boundary)
ps = geozoning:::interCB(cL[[8]],mapTest$step,mapTest$boundary,pG)#envelope is the whole map
sp::plot(pG)
sp::plot(ps,col="red",add=TRUE)
# not run
}
\keyword{internal}
