% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{linesC}
\alias{linesC}
\title{linesC}
\usage{
linesC(listContour, col = "blue")
}
\arguments{
\item{listContour}{list of contour lines}

\item{col}{line color}
}
\value{
an empty value
}
\description{
linesC
}
\details{
add contour Lines to plot
}
\examples{
data(mapTest)
cL=list()
cL=contourAuto(cL,mapTest$step,mapTest$xsize,mapTest$ysize,mapTest$krigGrid,c(5,7),mapTest$boundary)
plot(mapTest$boundary)
geozoning:::linesC(cL,col="black")
# not run
}
\keyword{internal}
