% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{linesSp}
\alias{linesSp}
\title{linesSp}
\usage{
linesSp(sp, k = 1, lty = 1, col = "red", lwd = 1)
}
\arguments{
\item{sp}{xxxx}

\item{k}{xxxx}

\item{lty}{xxxx}

\item{col}{xxxx}

\item{lwd}{xxxx}
}
\value{
a plot
}
\description{
linesSp
}
\details{
description, a paragraph
}
\examples{
# not run
}
