% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{listContourArea}
\alias{listContourArea}
\title{listContourArea}
\usage{
listContourArea(listContour)
}
\arguments{
\item{listContour}{xxxx}
}
\value{
a numeric?
}
\description{
listContourArea
}
\details{
description, a paragraph
}
\examples{
# not run
}
