% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{maxId}
\alias{maxId}
\title{maxId}
\usage{
maxId(Z)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}
}
\value{
a number
}
\description{
maxId
}
\details{
get highest number corresponding to a zone identifier in a zoning
}
\examples{
data(mapTest)
criti=correctionTree(c(0.4,0.5),mapTest,SAVE=TRUE)
Z=criti$zk[[2]][[1]]$zonePolygon
maxId(Z)
# not run
}
\keyword{internal}
