% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{meansdSimu}
\alias{meansdSimu}
\title{meansdSimu}
\usage{
meansdSimu(vseed = NULL, krig = 2)
}
\arguments{
\item{vseed}{list of simulation seeds}

\item{krig}{type of kriging (1-variogram model-based, 2-inverse distance-based)}
}
\value{
a matrix with as many rows as simulations, and 4 columns, the first two columns give mean and standard deviation of generated raw data, the last two columns give mean and standard deviation of kriged data
}
\description{
meansdSimu
}
\details{
computes mean and standard deviation of a set of map simulations
}
\examples{
\donttest{
 meansdSimu(c(1,2))
}
}
