% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{MeanVarWPts}
\alias{MeanVarWPts}
\title{MeanVarWPts}
\usage{
MeanVarWPts(map, zone, w = NULL)
}
\arguments{
\item{map}{xxxx}

\item{zone}{xxxx,}

\item{w}{default NULL}
}
\value{
a in a list
\describe{
\item{mean}{mean}
\item{var}{var}
}
}
\description{
MeanVarWPts
}
\details{
description, a paragraph
}
\examples{
# not run
}
