% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{modlm}
\alias{modlm}
\title{modlm}
\usage{
modlm(ptsp, Z)
}
\arguments{
\item{ptsp}{SpatialPointsDataFrame with data values}

\item{Z}{zoning (list of SpatialPolygons)}
}
\value{
the result of a call to lm (anova model with zone number as factor)
}
\description{
modlm
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
data(mapTest)
ptsp=mapTest$krigData
geozoning:::modlm(ptsp,Z)
# not run
}
\keyword{internal}
