% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{moveHoles}
\alias{moveHoles}
\title{moveHoles}
\usage{
moveHoles(zonePrinc, zoneSuppr)
}
\arguments{
\item{zonePrinc}{xxxx}

\item{zoneSuppr}{xxxx}
}
\value{
a ?
}
\description{
moveHoles
}
\details{
description, a paragraph
}
\examples{
# not run
}
