% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normDistMat}
\alias{normDistMat}
\title{normDistMat}
\usage{
normDistMat(matDistanceCorr, optiCrit)
}
\arguments{
\item{matDistanceCorr}{corrected distance matrix using pErr, result of calDistance}

\item{optiCrit}{criterion choice}
}
\value{
a normalized distance matrix
}
\description{
normDistMat
}
\details{
normalize distance matrix so that diagonal is equal to 1
}
\examples{
# load test map with simulated data
data(mapTest)
# load zoning results from test file
data(resZTest)
K=resZTest
resD = calDistance(typedist=1,mapTest$krigData,K$listZonePoint,K$zoneN,
        mapTest$krigSurfVoronoi,K$meanZone,pErr=0.9)
normDistMat(resD$matDistanceCorr,2)
}
