% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{normSize}
\alias{normSize}
\title{normSize}
\usage{
normSize(boundaryN, minSize, minSizeNG)
}
\arguments{
\item{boundaryN}{xxxx}

\item{minSize}{xxxx}

\item{minSizeNG}{xxxx}
}
\value{
a ?
}
\description{
normSize
}
\details{
description, a paragraph
}
\examples{
# not run
}
