% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{nPolyZone}
\alias{nPolyZone}
\title{nPolyZone}
\usage{
nPolyZone(Z, iC)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iC}{current zone number within Z}
}
\value{
the number of polygons within the current zone
}
\description{
nPolyZone
}
\details{
number of polygons in current zone
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
iC=1;print(paste(nPolyZone(Z,iC),"polygons in zone",iC))
}
