% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiRG.R
\name{optiRG}
\alias{optiRG}
\title{optiRG}
\usage{
optiRG(Z, K, map, iC, iZC, simplitol, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{K}{xxxx}

\item{map}{xxxx}

\item{iC}{xxxx}

\item{iZC}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
optiRG
}
\details{
description, a paragraph
}
\examples{
# not run
}
