% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotSp}
\alias{plotSp}
\title{plotSp}
\usage{
plotSp(sp, k = 1, xlim, ylim)
}
\arguments{
\item{sp}{xxxx}

\item{k}{xxxx}

\item{xlim}{xxxx}

\item{ylim}{xxxx}
}
\value{
a plot
}
\description{
plotSp
}
\details{
description, a paragraph
}
\examples{
# not run
}
