% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{plotVario}
\alias{plotVario}
\title{plotVario}
\usage{
plotVario(map, ylim = NULL)
}
\arguments{
\item{map}{object returned by function genMap}

\item{ylim}{range of y axis}
}
\value{
a plot
}
\description{
plotVario
}
\details{
plot empirical variogram for model and data in map (raw data plus kriged data)
}
\examples{
data(mapTest)
plotVario(mapTest)
}
