% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pointsSp}
\alias{pointsSp}
\title{pointsSp}
\usage{
pointsSp(sp, k = 1)
}
\arguments{
\item{sp}{xxxx}

\item{k}{xxxx}
}
\value{
a plot
}
\description{
pointsSp
}
\details{
description, a paragraph
}
\examples{
# not run
}
