% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{printInterZ}
\alias{printInterZ}
\title{printInterZ}
\usage{
printInterZ(Z, sp)
}
\arguments{
\item{Z}{list of zones, each zone is a SpatialPolygons}

\item{sp}{SpatialPolygons object}
}
\description{
printInterZ
}
\details{
checks intersection of sp and each element of Z
}
\examples{
\donttest{
data(mapTest)
data(resZTest)
K=resZTest
Z=K$zonePolygone
plotZ(Z,mapTest)
geozoning:::printInterZ(Z,Z[[1]])
}
}
\keyword{internal}
