% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{printZsurf}
\alias{printZsurf}
\title{printZsurf}
\usage{
printZsurf(Z, minSize = 0.012)
}
\arguments{
\item{Z}{xxxx}

\item{minSize}{xxxx}
}
\value{
a ?
}
\description{
printZsurf
}
\details{
description, a paragraph
}
\examples{
# not run
}
