% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ptInZone}
\alias{ptInZone}
\title{ptInZone}
\usage{
ptInZone(zone, pts, numpt)
}
\arguments{
\item{zone}{xxxx}

\item{pts}{xxxx}

\item{numpt}{xxxx}
}
\value{
a ?
}
\description{
ptInZone
}
\details{
description, a paragraph
}
\examples{
# not run
}
