% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcRandKmap.R
\name{ptNei}
\alias{ptNei}
\title{returns list of pt neigbors for each pt}
\usage{
ptNei(neighBool)
}
\arguments{
\item{neighBool}{numeric, boolean neighborhood matrix for pts}
}
\value{
a list of pt neigbors for each pt
}
\description{
returns list of pt neigbors for each pt
}
\examples{
# not run
}
