% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ptsInSp}
\alias{ptsInSp}
\title{ptInSp}
\usage{
ptsInSp(sp, pts, hole = FALSE)
}
\arguments{
\item{sp}{SpatialPolygons}

\item{pts}{data points}

\item{hole}{if TRUE also consider points in holes}
}
\value{
a data frame with data points within sp
}
\description{
ptInSp
}
\details{
finds data points in sp
}
\examples{
data(mapTest)
data(resZTest)
Z=resZTest$zonePolygone
ptsInSp(Z[[5]],mapTest$krigData) # 5 data points within zone 5

# not run
}
