% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randKmap.R
\name{randKmap}
\alias{randKmap}
\title{randKmap : todo title of the function in one line}
\usage{
randKmap(DataObj, seed, nPoints = 450, nPointsK = 2000, nSimuCond = 0,
  typeMod = "Gau", Vpsill = 5, Vrange = 0.2, Vmean = 8, Vnugget = 0,
  boundary = list(x = c(0, 0, 1, 1, 0), y = c(0, 1, 1, 0, 0)),
  manualBoundary = FALSE, krig = 1, disp = 0, FULL = FALSE)
}
\arguments{
\item{DataObj}{=NULL: simulated data with seed or = a data frame with real data}

\item{seed}{numeric, seed}

\item{nPoints}{numeric, number of points, default 450}

\item{nPointsK}{numeric, default 2000}

\item{nSimuCond}{numeric}

\item{typeMod}{character, model type}

\item{Vpsill}{numeric, default 5}

\item{Vrange}{numeric, default 0.2}

\item{Vmean}{numeric, default 8}

\item{Vnugget}{numeric, default 0}

\item{boundary}{list contains x and y}

\item{manualBoundary}{logical, default FALSE}

\item{krig}{numeric}

\item{disp}{numeric}

\item{FULL}{logical, if TRUE the returned list is complete}
}
\value{
a list
\describe{
\item{rawData}{rawData}
\item{step}{step}
\item{krigData}{krigData}
\item{krigGrid}{krigGrid}
\item{krigN}{krigN}
\item{krigSurfVoronoi}{krigSurfVoronoi}
\item{modelGen}{modelGen}
\item{modelVGM}{modelVGM}
\item{boundary}{boundary}
}
}
\description{
randKmap : todo title of the function in one line
}
\details{
description, a paragraph
}
\examples{
# not run
# map<-randKmap(DataObj,seed=seed,Vpsill=5,Vrange=0.2,Vnugget=0,Vmean=8,krig=1,disp=0,FULL=FALSE)

}
