% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randKmapGrid.R
\name{randKmapGrid}
\alias{randKmapGrid}
\title{randKmapGrid}
\usage{
randKmapGrid(DataObj, seed = NULL, nPointsK = 200, nSimuCond = 0,
  typeMod = "Gau", Vpsill = 5, Vrange = 0.2, Vmean = 8, Vang = 0,
  Vanis = 1, alphavario = NULL, boundary = list(x = c(0, 0, 1, 1, 0), y =
  c(0, 1, 1, 0, 0)), manualBoundary = FALSE, krig = 1, disp = 0,
  FULL = FALSE)
}
\arguments{
\item{DataObj}{xxxx}

\item{seed}{xxxx}

\item{nPointsK}{xxxx}

\item{nSimuCond}{xxxx}

\item{typeMod}{xxxx}

\item{Vpsill}{xxxx}

\item{Vrange}{xxxx}

\item{Vmean}{xxxx}

\item{Vang}{xxxx}

\item{Vanis}{xxxx}

\item{alphavario}{xxxx}

\item{boundary}{xxxx}

\item{manualBoundary}{xxxx}

\item{krig}{xxxx}

\item{disp}{xxxx}

\item{FULL}{xxxx}
}
\value{
a ?
}
\description{
randKmapGrid
}
\details{
description, a paragraph
}
\examples{
# not run
}
