% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readS.R
\name{readS}
\alias{readS}
\title{readS returns coords, ranges for x and y of a shapefile}
\usage{
readS(file, dir)
}
\arguments{
\item{file}{file name}

\item{dir}{directory}
}
\value{
a list with components
SpatialPolygonsDataFrame, ranges for x and y
}
\description{
readS returns coords, ranges for x and y of a shapefile
}
\details{
reads a polygon shp file in a directory and extracts coordinates and x and y ranges.
}
\examples{
\donttest{
#  readS was used to create the shape1 object in geozoning package
z=readS("Field_8_zones.shp",dir="../data/")
plot(z$sp)
}
}
