% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{remove1FromZ}
\alias{remove1FromZ}
\title{remove1FromZ}
\usage{
remove1FromZ(Z, iC, zoneN, simplitol = 0.001, disp = 0)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iC}{current zone index}

\item{zoneN}{zone neighborhood Logical matrix}

\item{simplitol}{tolerance for spatial polygons geometry simplification}

\item{disp}{0: no info, 1: some info}
}
\value{
a new zoning where current zone has been removed
}
\description{
remove1FromZ
}
\details{
description, a paragraph
}
\examples{
data(resZTest)
K=resZTest
Z=K$zonePolygone
plotZ(Z)
plotZ(remove1FromZ(Z,2,K$zoneN))
# not run
}
\keyword{internal}
