% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{removeFromZ}
\alias{removeFromZ}
\title{removeFromZ}
\usage{
removeFromZ(Z, zoneN, ptN, listZonePoint, data, simplitol, n = 1)
}
\arguments{
\item{Z}{xxxx}

\item{zoneN}{xxxx}

\item{ptN}{xxxx}

\item{listZonePoint}{xxxx}

\item{data}{xxxx}

\item{simplitol}{xxxx}

\item{n}{xxxx}
}
\value{
a ?
}
\description{
removeFromZ
}
\details{
description, a paragraph
}
\examples{
# not run
}
