% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveZK.R
\name{saveZK}
\alias{saveZK}
\title{saveZK}
\usage{
saveZK(map, K1, Z2, qProb, listOfZ, counter, crit, cost, costL, nz, mdist, pErr,
  optiCrit, simplitol)
}
\arguments{
\item{map}{xxxx}

\item{K1}{xxxx}

\item{Z2}{xxxx}

\item{qProb}{xxxx}

\item{listOfZ}{xxxx}

\item{counter}{xxxx}

\item{crit}{xxxx}

\item{cost}{xxxx}

\item{costL}{xxxx}

\item{nz}{xxxx}

\item{mdist}{xxxx}

\item{pErr}{xxxx}

\item{optiCrit}{xxxx}

\item{simplitol}{xxxx}
}
\value{
a ?
}
\description{
saveZK
}
\details{
description, a paragraph
}
\examples{
# not run
}
