% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{searchNODcrit}
\alias{searchNODcrit}
\title{searchNODcrit}
\usage{
searchNODcrit(qProb, le, zk, critere, cost, costL, nz)
}
\arguments{
\item{qProb}{xxxx}

\item{le}{xxxx}

\item{zk}{xxxx}

\item{critere}{xxxx}

\item{cost}{xxxx}

\item{costL}{xxxx}

\item{nz}{xxxx}
}
\value{
a ?
}
\description{
searchNODcrit
}
\details{
description, a paragraph
}
\examples{
# not run
}
