% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{setId}
\alias{setId}
\title{setId}
\usage{
setId(Z, iZ, id)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{iZ}{zone number}

\item{id}{zone identifier to assign}
}
\value{
a zoning geometry
}
\description{
setId
}
\details{
assign zone identifier in a zoning
}
\examples{
# not run, take a while...
#data(mapTest)
#criti=correctionTree(c(0.4,0.5),mapTest,SAVE=TRUE)
#Z=criti$zk[[2]][[1]]$zonePolygon
#Z1=setId(Z,4,"4")
}
