% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDistance.R
\name{Sigmai2}
\alias{Sigmai2}
\title{Sigmai2}
\usage{
Sigmai2(index, listZonePoint, tabVal, surfaceVoronoi, meanZone)
}
\arguments{
\item{index}{zone number}

\item{listZonePoint}{list of pts within each zone}

\item{tabVal}{SpatalPointsDataFrame holding data}

\item{surfaceVoronoi}{vector of Voronoi surfaces associated to data values}

\item{meanZone}{Zone mean values}
}
\value{
a list with components sigmai2 and SI
}
\description{
Sigmai2
}
\details{
computes mean (weighted) variance and Voronoi area for zone
}
\examples{
data(mapTest)
data(resZTest)
K=resZTest
Sigmai2(5,K$listZonePoint,mapTest$krigData,mapTest$krigSurfVoronoi,K$meanZone)
}
