% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{spnorm}
\alias{spnorm}
\title{spnorm}
\usage{
spnorm(sp, boundary)
}
\arguments{
\item{sp}{object of class Polygons}

\item{boundary}{list with x and y components, used to normalize sp}
}
\value{
a list with components
\describe{
\item{pn}{normalized Polygon}
\item{boundaryn}{normalized boundary}
}
}
\description{
spnorm
}
\details{
normalize Polygon according to border limits
}
\examples{
z=geozoning::shape1
bb=list(x=z@bbox[1,],y=z@bbox[2,])
p=z@polygons
p1=p[[1]]
P1=p1@Polygons[[1]]
NP1=geozoning:::spnorm(P1,bb)$pn
Nbb=geozoning:::spnorm(P1,bb)$boundaryn
plot(NP1@coords,xlim=Nbb$x,ylim=Nbb$y)
}
\keyword{internal}
