% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilZ.R
\name{testInterSpe}
\alias{testInterSpe}
\title{testInterSpe}
\usage{
testInterSpe(Z, i1, i2)
}
\arguments{
\item{Z}{zoning geometry (list of SpatialPolygons)}

\item{i1}{first zone number}

\item{i2}{second zone number}
}
\value{
a Logical value, TRUE if there is an intersection, FALSE if not.
}
\description{
testInterSpe
}
\details{
checks if 2 zones in a zoning share somme common part (using gOverlaps)
}
\examples{
\donttest{
data(mapTest)
qProb=c(0.2,0.5)
ZK = initialZoning(qProb, mapTest)
K=ZK$resZ
Z=K$zonePolygone
plotZ(Z)
Z58=rgeos::gConvexHull(rgeos::gUnion(Z[[8]],Z[[5]]))
Z[[length(Z)+1]]=Z58 # add new zone to zoning
plotZ(Z)
geozoning:::testInterSpe(Z,6,length(Z))
}
}
\keyword{internal}
