% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCalNei.R
\name{wMean}
\alias{wMean}
\title{wMean}
\usage{
wMean(type, listZonePoint, surfVoronoi, data)
}
\arguments{
\item{type}{xxxx}

\item{listZonePoint}{xxxx}

\item{surfVoronoi}{xxxx}

\item{data}{xxxx}
}
\value{
a ?
}
\description{
wMean
}
\details{
description, a paragraph
}
\examples{
# not run
}
