% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{zoneAssign}
\alias{zoneAssign}
\title{zoneAssign}
\usage{
zoneAssign(tab, Z)
}
\arguments{
\item{tab}{data frame with data values}

\item{Z}{zoning object}
}
\value{
a list of data points within each zone
}
\description{
zoneAssign
}
\details{
assigns points to zones
}
\examples{
data(mapTest)
ZK=initialZoning(qProb=c(0.4,0.7),mapTest)
Z=ZK$resZ$zonePolygone
listZpts=zoneAssign(mapTest$krigData,Z)
#identical to ZK$resZ$listZonePoint
listZptsRaw=zoneAssign(mapTest$rawData,Z)
plotZ(Z)
points(mapTest$rawData[listZptsRaw[[1]],],col="blue") # add raw data for zone 1
}
