% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{zoneFusion2}
\alias{zoneFusion2}
\title{zoneFusion2}
\usage{
zoneFusion2(zoneMain, zoneSuppr, simplitol)
}
\arguments{
\item{zoneMain}{xxxx}

\item{zoneSuppr}{xxxx}

\item{simplitol}{xxxx}
}
\value{
a ?
}
\description{
zoneFusion2
}
\details{
description, a paragraph
}
\examples{
# not run
}
