% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCleaning.R
\name{zoneFusion4}
\alias{zoneFusion4}
\title{zoneFusion4}
\usage{
zoneFusion4(Z, iSmall, iBig, map, simplitol, disp = 0)
}
\arguments{
\item{Z}{xxxx}

\item{iSmall}{xxxx}

\item{iBig}{xxxx}

\item{map}{xxxx}

\item{simplitol}{xxxx}

\item{disp}{xxxx}
}
\value{
a ?
}
\description{
zoneFusion4
}
\details{
description, a paragraph
}
\examples{
# not run
}
