% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcZoning.R
\name{zoneGeneration}
\alias{zoneGeneration}
\title{zoneGeneration}
\usage{
zoneGeneration(map, qProb = c(0.25, 0.75), GridData = FALSE)
}
\arguments{
\item{map}{object returned by function genMap}

\item{qProb}{probability vector used to generate quantile values}

\item{GridData}{logical value indicating if data are already on a regular grid (no kriging in that case)}
}
\value{
a list of zones, each zone is a SpatialPolygons
}
\description{
zoneGeneration
}
\details{
Generates zones from map data using quantile values associated to given probabilities
}
\examples{
data(mapTest)
Z=zoneGeneration(mapTest)
}
